% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{consume}
\alias{consume}
\title{Consume out of cash on hand}
\usage{
consume(Y, A, P, x_grid, cr, t, R)
}
\arguments{
\item{Y}{labor income}

\item{A}{assets}

\item{P}{permanent income for scaling cash on hand}

\item{cr}{optimal consumption rule
solved for with consumption_rule function}

\item{R}{gross return on assets}
}
\description{
Consume out of cash on hand
}
\details{
This computes consumption given income
assets and return on assets by first computing cash
on hand and then passing to the optimal consumption
rule.
}
