% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{consumption_path}
\alias{consumption_path}
\title{Get full consumption rule}
\usage{
consumption_path(
  x,
  G,
  sigma_n,
  sigma_u,
  gamma_0,
  gamma_1,
  R,
  p_noinc,
  beta,
  rho
)
}
\arguments{
\item{x}{cash on hand grid}

\item{G}{permanent income growth rate}

\item{sigma_n}{standard deviation of log permanent income}

\item{sigma_u}{standard deviation of log transitory income}

\item{gamma_0}{intercept of retirement MPC}

\item{gamma_1}{slope of retirement MPC in cash on hand}

\item{R}{risk free rate}

\item{beta}{time discount factor}

\item{T}{number of years}
}
\description{
Get full consumption rule
}
