% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{consumption_profile}
\alias{consumption_profile}
\title{Consumption Profile}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 40 rows and 3 columns.
}
\usage{
consumption_profile
}
\description{
Consumption Profile
}
\details{
The smoothed consumption profile data
used in the original paper. Taken from POG's website:
https://sites.google.com/view/pgourinchas/home/research/refereed-publications?authuser=0
}
\keyword{datasets}
