% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{create_coh_grid}
\alias{create_coh_grid}
\title{Create cash on hand grid}
\usage{
create_coh_grid(x_max, x_int, n_points)
}
\arguments{
\item{x_max}{max of the cash on hand grid, 40 in the paper}

\item{x_int}{internal stopping point for finer grid}

\item{n_points}{total number of points in the cash on hand grid}
}
\description{
Create cash on hand grid
}
\details{
this just follows the paper, a log spaced grid might make
more sense
}
