% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{draw_bernoulli}
\alias{draw_bernoulli}
\title{Draw a random matrix of bernoullis}
\usage{
draw_bernoulli(p, N, T)
}
\arguments{
\item{p}{probability of a 0}

\item{N}{number of rows}

\item{T}{number of cols}
}
\description{
Draw a random matrix of bernoullis
}
\details{
the probability is of a _zero_ not a one
because we are generating the probability of zero
income and it makes it easier, even though this is
non-standard.
}
