% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{endLog}
\alias{endLog}
\title{Stop logging started by startLog()}
\usage{
endLog(conn)
}
\arguments{
\item{conn}{Connection object returned by \code{startLog()}.}
}
\description{
Stop redirecting the stdout and stderr to a log file and close the connection returned by
\code{startLog}.
}
\examples{
\dontrun{
LOG <- startLog("mylog.txt", tee = interactive(), append = TRUE)
# ... do work ...
endLog(LOG)
}
}
