% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculus.R
\name{flatten}
\alias{flatten}
\title{Flattens sparse matrix in column-by-column order
This flattening procedure must be exactly the same as performed by RTMB Tape class, method $jacfun(sparse = TRUE),
see https://github.com/kaskr/RTMB/issues/85}
\usage{
flatten(S)
}
\description{
Flattens sparse matrix in column-by-column order
This flattening procedure must be exactly the same as performed by RTMB Tape class, method $jacfun(sparse = TRUE),
see https://github.com/kaskr/RTMB/issues/85
}
