% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{fn_classify_allele_frequencies}
\alias{fn_classify_allele_frequencies}
\title{Classify or bin allele frequencies into genotype classes.
This uses the ploidy level of the species to define these genotype classes,
e.g. for diploids we expect 3 genotype classes - AA, AB/BA, and BB,
while for tetraploids we expect 5 genotype classes - AAAA, AAAB, AABB, ABBB, and BBBB.}
\usage{
fn_classify_allele_frequencies(G, ploidy = 2, verbose = FALSE)
}
\arguments{
\item{G}{numeric n samples x p loci-alleles matrix of allele frequencies with non-null row and column names.
Row names can be any string of characters which identify the sample or entry or pool names.
Column names need to be tab-delimited, where first element refers to the chromosome or scaffold name,
the second should be numeric which refers to the position in the chromosome/scaffold, and
subsequent elements are optional which may refer to the allele identifier and other identifiers.}

\item{ploidy}{ploidy level which can refer to the number of haploid genomes to simulate pools (Default=2)}

\item{verbose}{show genotype binning or classification messages? (Default=FALSE)}
}
\value{
Ok: named n samples x p loci-alleles matrix of numeric genotype classes
Err: gpError
}
\description{
Classify or bin allele frequencies into genotype classes.
This uses the ploidy level of the species to define these genotype classes,
e.g. for diploids we expect 3 genotype classes - AA, AB/BA, and BB,
while for tetraploids we expect 5 genotype classes - AAAA, AAAB, AABB, ABBB, and BBBB.
}
\examples{
ploidy = 4
G = simquantgen::fn_simulate_genotypes(ploidy=ploidy, verbose=TRUE)
G_classes = fn_classify_allele_frequencies(G=G, ploidy=ploidy, verbose=TRUE)
}
