% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_validation.R
\name{fn_cross_validation_across_populations_pairwise}
\alias{fn_cross_validation_across_populations_pairwise}
\title{Pairwise-population cross-validation}
\usage{
fn_cross_validation_across_populations_pairwise(
  list_merged,
  vec_models_to_test = c("ridge", "lasso", "elastic_net", "Bayes_A", "Bayes_B",
    "Bayes_C", "gBLUP"),
  bool_parallel = TRUE,
  max_mem_Gb = 15,
  n_threads = 2,
  dir_output = NULL,
  verbose = FALSE
)
}
\arguments{
\item{list_merged}{list of merged genotype matrix, and phenotype vector, as well as an optional covariate matrix
\itemize{
\item $G: numeric n samples x p loci-alleles matrix of allele frequencies with non-null row and column names.
Row names can be any string of characters which identify the sample or entry or pool names.
Column names need to be tab-delimited, where first element refers to the chromosome or scaffold name,
the second should be numeric which refers to the position in the chromosome/scaffold, and
subsequent elements are optional which may refer to the allele identifier and other identifiers.
\item $list_pheno:
\itemize{
\item $y: named vector of numeric phenotype data
\item $pop: population or groupings corresponding to each element of y
\item $trait_name: name of the trait
}
\item $COVAR: numeric n samples x k covariates matrix with non-null row and column names.
}}

\item{vec_models_to_test}{genomic prediction models to use (uses all the models below by default). Please choose one or more from:
\itemize{
\item "ridge": \eqn{Cost_{ridge} = \Sigma(y - X\beta)^2 + \lambda\Sigma\beta^2$, where $\hat{\beta} = {(X^TX + \lambda I)^{-1} X^Ty}}
https://en.wikipedia.org/wiki/Ridge_regression
\item "lasso": \eqn{Cost_{lasso} = \Sigma(y - X\beta)^2 + \lambda\Sigma|\beta|}
https://en.wikipedia.org/wiki/Lasso_(statistics)
\item "elastic_net": \eqn{Cost_{ridge} = \Sigma(y - X\beta)^2 + (1-\alpha)\lambda\Sigma\beta^2 + \alpha\lambda\Sigma|\beta|}
https://en.wikipedia.org/wiki/Elastic_net_regularization
\item "Bayes_A": scaled t-distributed effects
https://cran.r-hub.io/web/packages/BGLR/vignettes/BGLR-extdoc.pdf
\item "Bayes_B": scaled t-distributed effects with probability \eqn{\pi}; and zero effects with probability \eqn{1-\pi},
where \eqn{\pi \sim \beta(\theta_1, \theta_2)}.
https://cran.r-hub.io/web/packages/BGLR/vignettes/BGLR-extdoc.pdf
\item "Bayes_C": normally distributed effects (\eqn{N(0, \sigma^2_{\beta})}) with probability \eqn{\pi};
and zero effects with probability \eqn{1-\pi}, where \eqn{\pi \sim \beta(\theta_1, \theta_2)}.
https://cran.r-hub.io/web/packages/BGLR/vignettes/BGLR-extdoc.pdf
\item "gBLUP": genotype best linear unbiased prediction (gBLUP) using genomic relationship matrix to predict missing breeding values
via Direct-Inversion Newton-Raphson or Average Information
using the sommer R package (https://www.ncbi.nlm.nih.gov/pmc/articles/PMC4894563/).
https://link.springer.com/protocol/10.1007/978-1-62703-447-0_13
}}

\item{bool_parallel}{perform multi-threaded cross-validation (Default=TRUE)}

\item{max_mem_Gb}{maximum memory in gigabytes available for computation (Default=15)}

\item{n_threads}{total number of computing threads available (Default=2)}

\item{dir_output}{output directory where temporary text and Rds files will be saved into (Default=NULL)}

\item{verbose}{show bulk across population cross-validation messages? (Default=FALSE)}
}
\value{
\itemize{
\item Ok: filename of temporary Rds file containing a 2-element list:
\itemize{
\item $METRICS_ACROSS_POP_PAIRWISE:
\itemize{
\item $rep: replication number
\item $fold: fold number
\item $model: genomic prediction model name
\item $pop_training: population/s used in the training set (separated by commas if more than 1)
\item $pop_validation: population/s used in the validation set (separated by commas if more than 1)
\item $n_training: number of samples/entries/pools in the training set
\item $n_validation: number of samples/entries/pools in the validation set
\item $duration_mins: time taken in minutes to fit the genomic prediction model and assess the prediction accuracies
\item $n_non_zero: number of non-zero estimated effects (effects greater than machine epsilon ~2.2e-16)
\item $n_total_features: total number of zero and non-zero effects, i.e. total number of SNPs/alleles/features
\item $mbe: mean bias error
\item $mae: mean absolute error
\item $rmse: root mean squared error
\item $r2: coefficient of determination
\item $corr: Pearson's product moment correlation
\item $power_t10: fraction of observed top 10 phenotype values correctly predicted
\item $power_b10: fraction of observed bottom 10 phenotype values correctly predicted
\item $var_pred: variance of predicted phenotype values (estimator of additive genetic variance)
\item $var_true: variance of observed phenotype values (estimator of total phenotypic variance)
\item $h2: narrow-sense heritability estimate
}
\item $YPRED_ACROSS_POP_PAIRWISE:
\itemize{
\item $rep: replication number
\item $fold: fold number
\item $model: genomic prediction model name
\item $pop_training: population/s used in the training set (separated by commas if more than 1)
\item $id: names of the samples/entries/pools,
\item $pop_validation: population from which the sample/entry/pool belongs to
\item $y_true: observed phenotype values
\item $y_pred: predicted phenotype values
}
}
\item Err: gpError
}
}
\description{
Pairwise-population cross-validation
}
\examples{
list_sim = fn_simulate_data(n_pop=3, verbose=TRUE)
G = fn_load_genotype(fname_geno=list_sim$fname_geno_vcf)
list_pheno = fn_load_phenotype(fname_pheno=list_sim$fname_pheno_tsv)
COVAR = G \%*\% t(G)
list_merged = fn_merge_genotype_and_phenotype(G=G, list_pheno=list_pheno, COVAR=COVAR, verbose=TRUE)
fname_across_pairwise_Rds = fn_cross_validation_across_populations_pairwise(list_merged, 
 vec_models_to_test=c("ridge","lasso"), verbose=TRUE)
list_across_pairwise = readRDS(fname_across_pairwise_Rds)
head(list_across_pairwise$METRICS_ACROSS_POP_PAIRWISE)
head(list_across_pairwise$YPRED_ACROSS_POP_PAIRWISE)
}
