% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_validation.R
\name{fn_cross_validation_preparation}
\alias{fn_cross_validation_preparation}
\title{Define the type of cross-validation scheme including the models, partitioning, and shuffling,
as well as estimate the maximum number of threads which can be used in parallel without memory errors.}
\usage{
fn_cross_validation_preparation(
  list_merged,
  cv_type = 1,
  n_folds = 10,
  n_reps = 10,
  vec_models_to_test = c("ridge", "lasso", "elastic_net", "Bayes_A", "Bayes_B",
    "Bayes_C", "gBLUP"),
  max_mem_Gb = 15,
  verbose = FALSE
)
}
\arguments{
\item{list_merged}{list of merged genotype matrix, and phenotype vector, as well as an optional covariate matrix
\itemize{
\item $G: numeric n samples x p loci-alleles matrix of allele frequencies with non-null row and column names.
Row names can be any string of characters which identify the sample or entry or pool names.
Column names need to be tab-delimited, where first element refers to the chromosome or scaffold name,
the second should be numeric which refers to the position in the chromosome/scaffold, and
subsequent elements are optional which may refer to the allele identifier and other identifiers.
\item $list_pheno:
\itemize{
\item $y: named vector of numeric phenotype data
\item $pop: population or groupings corresponding to each element of y
\item $trait_name: name of the trait
}
\item $COVAR: numeric n samples x k covariates matrix with non-null row and column names.
}}

\item{cv_type}{(Default=1)
\itemize{
\item 1: k-fold cross-validation across all samples/entries/pools regardless of their groupings
\item 2: pairwise-population cross-validation, e.g. training on population A and validation on population B
\item 3: leave-one-population-out cross-validation, e.g. training on populations 1 to 9 and validation on population 10
}}

\item{n_folds}{if cv_type=1, i.e. k-fold cross-validation is desired, then this defines the number of partitions or folds.
However, if the resulting size per fold is less than 2, then this return an error.
Additionally, this parameter is ignored is cv_type=2 or cv_type=3 (Default=10).}

\item{n_reps}{if cv_type=1, i.e. k-fold cross-validation is desired, then this defines the number of random shuffling
for repeated k-fold cross-validation.
Similar to n_folds, this parameter is ignored is cv_type=2 or cv_type=3 (Default=10).}

\item{vec_models_to_test}{genomic prediction models to use (uses all the models below by default). Please choose one or more from:
\itemize{
\item "ridge": \eqn{Cost_{ridge} = \Sigma(y - X\beta)^2 + \lambda\Sigma\beta^2$, where $\hat{\beta} = {(X^TX + \lambda I)^{-1} X^Ty}}
https://en.wikipedia.org/wiki/Ridge_regression
\item "lasso": \eqn{Cost_{lasso} = \Sigma(y - X\beta)^2 + \lambda\Sigma|\beta|}
https://en.wikipedia.org/wiki/Lasso_(statistics)
\item "elastic_net": \eqn{Cost_{ridge} = \Sigma(y - X\beta)^2 + (1-\alpha)\lambda\Sigma\beta^2 + \alpha\lambda\Sigma|\beta|}
https://en.wikipedia.org/wiki/Elastic_net_regularization
\item "Bayes_A": scaled t-distributed effects
https://cran.r-hub.io/web/packages/BGLR/vignettes/BGLR-extdoc.pdf
\item "Bayes_B": scaled t-distributed effects with probability \eqn{\pi}; and zero effects with probability \eqn{1-\pi},
where \eqn{\pi \sim \beta(\theta_1, \theta_2)}.
https://cran.r-hub.io/web/packages/BGLR/vignettes/BGLR-extdoc.pdf
\item "Bayes_C": normally distributed effects (\eqn{N(0, \sigma^2_{\beta})}) with probability \eqn{\pi};
and zero effects with probability \eqn{1-\pi}, where \eqn{\pi \sim \beta(\theta_1, \theta_2)}.
https://cran.r-hub.io/web/packages/BGLR/vignettes/BGLR-extdoc.pdf
\item "gBLUP": genotype best linear unbiased prediction (gBLUP) using genomic relationship matrix to predict missing breeding values
via Direct-Inversion Newton-Raphson or Average Information
using the sommer R package (https://www.ncbi.nlm.nih.gov/pmc/articles/PMC4894563/).
https://link.springer.com/protocol/10.1007/978-1-62703-447-0_13
}}

\item{max_mem_Gb}{maximum memory in gigabytes available for computation (Default=15)}

\item{verbose}{show cross-validation parameter preparation messages? (Default=FALSE)}
}
\value{
\itemize{
\item Ok:
\item $df_params: data frame containing all possible or just a defined non-exhaustive combinations of:
\itemize{
\item $rep: replication number
\item $fold: fold number
\item $model: model name
}
\item $mat_idx_shuffle numeric: n sample x r replications matrix of sample/entry/pool index shuffling where
each column refer to a random shuffling of samples/entry/pool from which the identities of the
training and validation sets will be sourced from
\item $vec_set_partition_groupings vector of numeric partitioning indexes where each index refer to the
fold which will serve as the validation population
\item Err: gpError
}
}
\description{
Define the type of cross-validation scheme including the models, partitioning, and shuffling,
as well as estimate the maximum number of threads which can be used in parallel without memory errors.
}
\examples{
list_sim = fn_simulate_data(n_pop=3, verbose=TRUE)
G = fn_load_genotype(fname_geno=list_sim$fname_geno_vcf)
list_pheno = fn_load_phenotype(fname_pheno=list_sim$fname_pheno_tsv)
COVAR = G \%*\% t(G)
list_merged = fn_merge_genotype_and_phenotype(G=G, list_pheno=list_pheno, COVAR=COVAR, verbose=TRUE)
list_cv_params = fn_cross_validation_preparation(list_merged, cv_type=1)
list_cv_params = fn_cross_validation_preparation(list_merged, cv_type=3)
list_cv_params = fn_cross_validation_preparation(list_merged, cv_type=2)
list_merged$list_pheno$pop = rep(c("popA", "popB"), times=length(list_merged$list_pheno$pop)/2)
list_cv_params = fn_cross_validation_preparation(list_merged, cv_type=2)
}
