% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{fn_estimate_memory_footprint}
\alias{fn_estimate_memory_footprint}
\title{Estimate memory usage for parallel replicated k-fold cross validation of multiple genomic prediction models}
\usage{
fn_estimate_memory_footprint(
  X,
  n_models = 7,
  n_folds = 10,
  n_reps = 10,
  memory_requested_Gb = 400,
  memory_multiplier = 40,
  verbose = FALSE
)
}
\arguments{
\item{X}{numeric matrix resulting from the merging of the genotype matrix, phenotype vector and covariate matrix}

\item{n_models}{number of genomic prediction models to fit. Note that Bayesian and gBLUP models are more memory-intensive than penalised regression ones. (Default=7)}

\item{n_folds}{number of cross-validation folds (Default=10)}

\item{n_reps}{number of cross-validation replication (Default=10)}

\item{memory_requested_Gb}{memory requested or available for use (Default=400)}

\item{memory_multiplier}{estimated memory usage multiplier (Default=40)}

\item{verbose}{show memory usage estimation messages? (Default=FALSE)}
}
\value{
Ok:
$size_X: memory used for a single merge genotype-phenotype-covariate dataset
$size_total: total memory required for parallel computations across models, folds and replications
$n_threads: recommended and estimated maximum number of threads to use to prevent out-of-memory (OOM) error
Err: gpError
}
\description{
Estimate memory usage for parallel replicated k-fold cross validation of multiple genomic prediction models
}
