% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{fn_G_extract_names}
\alias{fn_G_extract_names}
\title{Extract names of samples/entries/pools and loci from a numeric (allele frequency) or non-numeric (genotype classes) matrix}
\usage{
fn_G_extract_names(mat_genotypes, verbose = FALSE)
}
\arguments{
\item{mat_genotypes}{n samples x p loci matrix of:
allele frequencies (numeric ranging from 0 to 1) or
genotype classes (e.g. "AA", "AB", "AAB", "AAAB")
with non-null row and column names.
Row names can be any string of characters which identify the sample or entry or pool names.
Column names need to be tab-delimited, where first element refers to the chromosome or scaffold name,
the second should be numeric which refers to the position in the chromosome/scaffold, and
subsequent elements are optional which may refer to the allele identifier and other identifiers.}

\item{verbose}{show names extraction messages? (Default=FALSE)}
}
\value{
Ok:
$vec_ids: vector of sample/entry/pool names (same order as in G)
$vec_chr: vector of chromosome or scaffold names (same order as in G)
$vec_pos: vector of numeric positions per chromosome (same order as in G)
$vec_all: vector of allele names (same order as in G)
Err: grError
}
\description{
Extract names of samples/entries/pools and loci from a numeric (allele frequency) or non-numeric (genotype classes) matrix
}
\examples{
mat_genotypes = simquantgen::fn_simulate_genotypes(verbose=TRUE)
list_ids_chr_pos_all = fn_G_extract_names(mat_genotypes=mat_genotypes, verbose=TRUE)
}
