% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{fn_G_numeric_to_non_numeric}
\alias{fn_G_numeric_to_non_numeric}
\title{Convert a numeric (allele frequency) genotype matrix into a non-numeric (genotype classes) genotype matrix}
\usage{
fn_G_numeric_to_non_numeric(G, ploidy = 2, verbose = FALSE)
}
\arguments{
\item{G}{n samples x p loci matrix of allele frequencies (numeric ranging from 0 to 1) with non-null row and column names.
Row names can be any string of characters which identify the sample or entry or pool names.
Column names need to be tab-delimited, where first element refers to the chromosome or scaffold name,
the second should be numeric which refers to the position in the chromosome/scaffold, and
subsequent elements are optional which may refer to the allele identifier and other identifiers.}

\item{ploidy}{ploidy level which can refer to the number of haploid genomes to simulate pools (Default=2)}

\item{verbose}{show numeric to non-numeric genotype data conversion messages? (Default=FALSE)}
}
\value{
Ok: n samples x p loci matrix of genotype classes
Err: grError
}
\description{
Convert a numeric (allele frequency) genotype matrix into a non-numeric (genotype classes) genotype matrix
}
\examples{
ploidy = 42
G_numeric = simquantgen::fn_simulate_genotypes(ploidy=ploidy, n_alleles=52, verbose=TRUE)
G_non_numeric = fn_G_numeric_to_non_numeric(G=G_numeric, ploidy=ploidy, verbose=TRUE)
}
