% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{fn_load_genotype}
\alias{fn_load_genotype}
\title{Load genotype data as an allele frequency matrix
Load genotype data as an allele frequency matrix}
\usage{
fn_load_genotype(
  fname_geno,
  ploidy = NULL,
  retain_minus_one_alleles_per_locus = TRUE,
  min_depth = 0,
  max_depth = Inf,
  verbose = FALSE
)
}
\arguments{
\item{fname_geno}{filename of the genotype file
This may be in 1 of 3 formats:
- VCF format,
- allele frequency table saved as a tab-delimited file with a header line and the first 3 columns refer to the
chromosome (chr), position (pos), and allele (allele),
with subsequent columns referring to the allele frequencies of a sample, entry or pool.
Names of the samples, entries, or pools in the header line can be any unique string of characters.
- Rds file containing a single numeric n samples x p loci-alleles matrix of allele frequencies with non-null row and column names.
Row names can be any string of characters which identify the sample or entry or pool names.
Column names need to be tab-delimited, where first element refers to the chromosome or scaffold name,
the second should be numeric which refers to the position in the chromosome/scaffold, and
subsequent elements are optional which may refer to the allele identifier and other identifiers.}

\item{ploidy}{ploidy level which will generate genotype classes instead of continuous allele frequencies.
If NULL, then continuous allele frequencies and no binning or classification into genotype classes
will be performed (Default=NULL).}

\item{retain_minus_one_alleles_per_locus}{omit the alternative or trailing allele per locus? (Default=TRUE)}

\item{min_depth}{if input is a VCF file: minimum depth per locus beyond which will be set to missing data (Default=0)}

\item{max_depth}{if input is a VCF file: maximum depth per locus beyond which will be set to missing data (Default=Inf)}

\item{verbose}{show genotype loading messages? (Default=FALSE)}
}
\value{
Ok: numeric n samples x p loci-alleles matrix of allele frequencies with non-null row and column names.
Row names can be any string of characters which identify the sample or entry or pool names.
Column names need to be tab-delimited, where first element refers to the chromosome or scaffold name,
the second should be numeric which refers to the position in the chromosome/scaffold, and
subsequent elements are optional which may refer to the allele identifier and other identifiers.
Err: gpError
}
\description{
Load genotype data as an allele frequency matrix
Load genotype data as an allele frequency matrix
}
\examples{
list_sim = fn_simulate_data(verbose=TRUE, 
 save_geno_vcf=TRUE, save_geno_tsv=TRUE, save_geno_rds=TRUE, save_pheno_tsv=TRUE)
G_vcf = fn_load_genotype(fname_geno=list_sim$fname_geno_vcf, verbose=TRUE)
G_tsv = fn_load_genotype(fname_geno=list_sim$fname_geno_tsv, verbose=TRUE)
G_rds = fn_load_genotype(fname_geno=list_sim$fname_geno_rds, verbose=TRUE)
}
