% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{fn_load_phenotype}
\alias{fn_load_phenotype}
\title{Load phenotype data from a text-delimited file}
\usage{
fn_load_phenotype(
  fname_pheno,
  sep = "\\t",
  header = TRUE,
  idx_col_id = 1,
  idx_col_pop = 2,
  idx_col_y = 3,
  na.strings = c("", "-", "NA", "na", "NaN", "missing", "MISSING"),
  verbose = FALSE
)
}
\arguments{
\item{fname_pheno}{filname of the text-delimited phenotype data}

\item{sep}{column-delimiter in the phenotype file (Default="\t")}

\item{header}{does the phenotype file have a header line? (Default=TRUE)}

\item{idx_col_id}{which column correspond to the sample/entry/pool/genotype names? (Default=1)}

\item{idx_col_pop}{which column correspond to the population groupings? (Default=2)}

\item{idx_col_y}{which column correspond to the phenotype data? (Default=3)}

\item{na.strings}{vector of string corresponding to missing data in the phenotype column (Default=c("", "-", "NA", "na", "NaN", "missing", "MISSING"))}

\item{verbose}{show phenotype loading messages? (Default=FALSE)}
}
\value{
Ok:
$y: named vector of numeric phenotype data
$pop: population or groupings corresponding to each element of y
$trait_name: name of the trait
Err: gpError
}
\description{
Load phenotype data from a text-delimited file
}
\examples{
list_sim = fn_simulate_data(n_pop=3, verbose=TRUE)
list_pheno = fn_load_phenotype(fname_pheno=list_sim$fname_pheno_tsv)
}
