% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{fn_merge_genotype_and_phenotype}
\alias{fn_merge_genotype_and_phenotype}
\title{Merge genotype and phenotype data by their entry names, i.e. rownames for G and names for y.
All samples with genotype data will be included and samples without phenotype data
will be set to NA (all.x=TRUE). Samples with phenotype but without genotype data are omitted.}
\usage{
fn_merge_genotype_and_phenotype(G, list_pheno, COVAR = NULL, verbose = FALSE)
}
\arguments{
\item{G}{numeric n samples x p loci-alleles matrix of allele frequencies with non-null row and column names.
Row names can be any string of characters which identify the sample or entry or pool names.
Column names need to be tab-delimited, where first element refers to the chromosome or scaffold name,
the second should be numeric which refers to the position in the chromosome/scaffold, and
subsequent elements are optional which may refer to the allele identifier and other identifiers.}

\item{list_pheno}{list with 3 elements:
(1) $y: a named vector of numeric phenotype data,
(2) $pop: population or groupings corresponding to each element of y, and
(3) $trait_name: name of the trait.}

\item{COVAR}{numeric n samples x k covariates matrix with non-null row and column names.
The row names correspond to the sample or entry or pool names,
which fully or partially match with the row names in G and the names in list_pheno$y. (Default=NULL)}

\item{verbose}{show genotype, phenotype, and covariate merging messages? (Default=FALSE)}
}
\value{
Ok:
$G: numeric n samples x p loci-alleles matrix of allele frequencies with non-null row and column names.
Row names can be any string of characters which identify the sample or entry or pool names.
Column names need to be tab-delimited, where first element refers to the chromosome or scaffold name,
the second should be numeric which refers to the position in the chromosome/scaffold, and
subsequent elements are optional which may refer to the allele identifier and other identifiers.
$list_pheno:
$y: named vector of numeric phenotype data
$pop: population or groupings corresponding to each element of y
$trait_name: name of the trait
$COVAR: numeric n samples x k covariates matrix with non-null row and column names.
Err: gpError
}
\description{
Merge genotype and phenotype data by their entry names, i.e. rownames for G and names for y.
All samples with genotype data will be included and samples without phenotype data
will be set to NA (all.x=TRUE). Samples with phenotype but without genotype data are omitted.
}
\examples{
list_sim = fn_simulate_data(n_pop=3, verbose=TRUE)
G = fn_load_genotype(fname_geno=list_sim$fname_geno_vcf)
rownames(G)[1] = "entry_exclude_me"
rownames(G)[2] = "entry_exclude_me_too"
list_pheno = fn_load_phenotype(fname_pheno=list_sim$fname_pheno_tsv)
COVAR = matrix(stats::rnorm(n=(10*nrow(G))), nrow=nrow(G))
rownames(COVAR) = rownames(G); colnames(COVAR) = paste0("covariate_", 1:ncol(COVAR))
list_merged = fn_merge_genotype_and_phenotype(G=G, list_pheno=list_pheno, COVAR=COVAR, verbose=TRUE)
}
