% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{fn_merge_genotype_genotype}
\alias{fn_merge_genotype_genotype}
\title{Merge two genotypes matrices where if there are conflict:
\itemize{
\item data on the first matrix will be used,
\item data on the second matrix will be used, or
\item arithmetic mean between the two matrices will be used.
}}
\usage{
fn_merge_genotype_genotype(
  G1,
  G2,
  str_conflict_resolution = c("1-use-G1", "2-use-G2", "3-use_mean")[3],
  verbose = FALSE
)
}
\arguments{
\item{G1}{numeric n samples x p loci-alleles matrix of allele frequencies with non-null row and column names.
Row names can be any string of characters which identify the sample or entry or pool names.
Column names need to be tab-delimited, where first element refers to the chromosome or scaffold name,
the second should be numeric which refers to the position in the chromosome/scaffold, and
subsequent elements are optional which may refer to the allele identifier and other identifiers.}

\item{G2}{numeric n samples x p loci-alleles matrix of allele frequencies with non-null row and column names.
Row names can be any string of characters which identify the sample or entry or pool names.
Column names need to be tab-delimited, where first element refers to the chromosome or scaffold name,
the second should be numeric which refers to the position in the chromosome/scaffold, and
subsequent elements are optional which may refer to the allele identifier and other identifiers.}

\item{str_conflict_resolution}{conflict resolution mode. Use "1" to always use the genotype data from the
first matrix, "2" to to always use the data from the second matrix, and "3" to compute the arithmetic mean
between the two matrices.}

\item{verbose}{show genotype merging messages? (Default=FALSE)}
}
\value{
\itemize{
\item Ok: numeric n samples x p loci-alleles matrix of allele frequencies with non-null row and column names.
Row names can be any string of characters which identify the sample or entry or pool names.
Column names need to be tab-delimited, where first element refers to the chromosome or scaffold name,
the second should be numeric which refers to the position in the chromosome/scaffold, and
subsequent elements are optional which may refer to the allele identifier and other identifiers.
\item Err: gpError
}
}
\description{
Merge two genotypes matrices where if there are conflict:
\itemize{
\item data on the first matrix will be used,
\item data on the second matrix will be used, or
\item arithmetic mean between the two matrices will be used.
}
}
\examples{
list_sim = gp::fn_simulate_data(verbose=TRUE)
G = fn_load_genotype(fname_geno=list_sim$fname_geno_vcf)
G1 = G[1:ceiling(0.5*nrow(G)), ]
G2 = G[(ceiling(0.5*nrow(G))+1):nrow(G), ]
G_merged = fn_merge_genotype_genotype(G1=G1, G2=G2, verbose=TRUE)
}
