% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{fn_ridge}
\alias{fn_ridge}
\title{Ridge regression model (a.k.a. GBLUP; regularisation via Tikhonov regularisation; alpha=0)}
\usage{
fn_ridge(
  list_merged,
  vec_idx_training,
  vec_idx_validation,
  other_params = list(n_folds = 10),
  verbose = FALSE
)
}
\arguments{
\item{list_merged}{list of merged genotype matrix, and phenotype vector, as well as an optional covariate matrix
$G: numeric n samples x p loci-alleles matrix of allele frequencies with non-null row and column names.
Row names can be any string of characters which identify the sample or entry or pool names.
Column names need to be tab-delimited, where first element refers to the chromosome or scaffold name,
the second should be numeric which refers to the position in the chromosome/scaffold, and
subsequent elements are optional which may refer to the allele identifier and other identifiers.
$list_pheno:
$y: named vector of numeric phenotype data
$pop: population or groupings corresponding to each element of y
$trait_name: name of the trait
$COVAR: numeric n samples x k covariates matrix with non-null row and column names.}

\item{vec_idx_training}{vector of numeric indexes referring to the training set}

\item{vec_idx_validation}{vector of numeric indexes referring to the validation set}

\item{other_params}{list of additional parameters, specifically $n_folds which refers to the number of
internal cross-fold validation to find the optimum lambda at which the deviance is minimised (Default=list(n_folds=10))}

\item{verbose}{show ridge regression messages? (Default=FALSE)}
}
\value{
Ok:
$list_perf:
$mbe: mean bias error
$mae: mean absolute error
$rmse: root mean squared error
$r2: coefficient of determination
$corr: Pearson's product moment correlation
$power_t10: fraction of observed top 10 phenotype values correctly predicted
$power_b10: fraction of observed bottom 10 phenotype values correctly predicted
$var_pred: variance of predicted phenotype values (estimator of additive genetic variance)
$var_true: variance of observed phenotype values (estimator of total phenotypic variance)
$h2: narrow-sense heritability estimate
$df_y_validation: data frame of the validation set with names of the samples/entries/pools,
population, observed and predicted phenotypic values
$vec_effects: named numeric vector of estimated effects, where the names correspond to the
SNP/allele identity including chromosome/scaffold, position and optionally allele.
$n_non_zero: number of non-zero estimated effects (effects greater than machine epsilon ~2.2e-16)
Err: gpError
}
\description{
Ridge regression model (a.k.a. GBLUP; regularisation via Tikhonov regularisation; alpha=0)
}
\examples{
list_sim = fn_simulate_data(n_pop=3, verbose=TRUE)
G = fn_load_genotype(fname_geno=list_sim$fname_geno_vcf)
list_pheno = fn_load_phenotype(fname_pheno=list_sim$fname_pheno_tsv)
list_merged = fn_merge_genotype_and_phenotype(G=G, list_pheno=list_pheno, verbose=TRUE)
n = nrow(list_merged$G)
vec_idx_training = sample(c(1:n), floor(n/2))
vec_idx_validation = c(1:n)[!(c(1:n) \%in\% vec_idx_training)]
list_ols = fn_ols(list_merged, vec_idx_training, vec_idx_validation, verbose=TRUE)
}
