% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{fn_save_phenotype}
\alias{fn_save_phenotype}
\title{Save phenotype data into a delimited file
with 3 columns referring to the sample ID, population ID, and trait values}
\usage{
fn_save_phenotype(list_pheno, fname, sep = "\\t", verbose = FALSE)
}
\arguments{
\item{list_pheno}{list with 3 elements:
(1) $y: a named vector of numeric phenotype data,
(2) $pop: population or groupings corresponding to each element of y, and
(3) $trait_name: name of the trait.}

\item{fname}{file name of the output file}

\item{sep}{delimited of the output file (Default="\\t")}

\item{verbose}{show phenotype filtering messages? (Default=FALSE)}
}
\value{
\itemize{
\item Ok: file name of the output file
\item Err: gpError
}
}
\description{
Save phenotype data into a delimited file
with 3 columns referring to the sample ID, population ID, and trait values
}
\examples{
list_sim = fn_simulate_data(n_pop=3, verbose=TRUE)
list_pheno = fn_load_phenotype(fname_pheno=list_sim$fname_pheno_tsv)
fn_save_phenotype(list_pheno, fname=tempfile(fileext=".tsv"), sep="\t", verbose=TRUE)
fn_save_phenotype(list_pheno, fname=tempfile(fileext=".csv"), sep=",", verbose=TRUE)
}
