% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{fn_simulate_data}
\alias{fn_simulate_data}
\title{Simple wrapper of simquantgen simulation of 10 standard normally distributed QTL additive effects
on 5-chromosome genome and a single trait at 50\% heritability}
\usage{
fn_simulate_data(
  n = 100,
  l = 1000,
  ploidy = 2,
  n_alleles = 2,
  min_depth = 5,
  max_depth = 1000,
  seed = 12345,
  save_pheno_tsv = TRUE,
  save_geno_vcf = TRUE,
  save_geno_tsv = FALSE,
  save_geno_rds = FALSE,
  non_numeric_Rds = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{n}{number of samples \link{Default=100}}

\item{l}{number of loci \link{Default=1000}}

\item{ploidy}{ploidy level which can refer to the number of haploid genomes to simulate pools \link{Default=2}}

\item{n_alleles}{macimum number of alleles per locus \link{Default=2}}

\item{min_depth}{minimum depth per locus \link{Default=5}}

\item{max_depth}{maximum depth per locus \link{Default=5000}}

\item{seed}{randomisation seed for replicability \link{Default=12345}}

\item{save_pheno_tsv}{save the phenotype data as a tab-delimited file? \link{Default=TRUE}}

\item{save_geno_vcf}{save the genotype data as a vcf file? \link{Default=TRUE}}

\item{save_geno_tsv}{save the genotype data as a tab-delimited allele frequency table file? \link{Default=FALSE}}

\item{save_geno_rds}{save the named genotype matrix as an Rds file? \link{Default=FALSE}}

\item{non_numeric_Rds}{save non-numeric Rds genotype file? \link{Default=FALSE}}

\item{verbose}{show genotype and phenotype data simulation messages? \link{Default=FALSE}}
}
\value{
vcf: simulated genotype data as a vcfR object
df: simulated phenotype data as a data frame
fname_geno_vcf: filename of the simulated genotype data as a vcf file
fname_geno_tsv: filename of the simulated genotype data as a tab-delimited allele frequency table file
fname_geno_rds: filename of the simulated named genotype matrix as an Rds file
fname_pheno_tsv: filename of the simulated phenotype data as a tab-delimited file
}
\description{
Simple wrapper of simquantgen simulation of 10 standard normally distributed QTL additive effects
on 5-chromosome genome and a single trait at 50\% heritability
}
\examples{
list_sim = fn_simulate_data(verbose=TRUE, save_geno_vcf=TRUE, save_geno_tsv=TRUE, save_geno_rds=TRUE, save_pheno_tsv=TRUE)
}
