% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{fn_subset_merged_genotype_and_phenotype}
\alias{fn_subset_merged_genotype_and_phenotype}
\title{Subset the list of merged genotype and phenotype data using a vector of sample/entry/pool indexes.}
\usage{
fn_subset_merged_genotype_and_phenotype(list_merged, vec_idx, verbose = FALSE)
}
\arguments{
\item{list_merged}{list of merged genotype matrix, and phenotype vector, as well as an optional covariate matrix
+ $G: numeric n samples x p loci-alleles matrix of allele frequencies with non-null row and column names.
Row names can be any string of characters which identify the sample or entry or pool names.
Column names need to be tab-delimited, where first element refers to the chromosome or scaffold name,
the second should be numeric which refers to the position in the chromosome/scaffold, and
subsequent elements are optional which may refer to the allele identifier and other identifiers.
+ $list_pheno:
$y: named vector of numeric phenotype data
$pop: population or groupings corresponding to each element of y
$trait_name: name of the trait
+ $COVAR: numeric n samples x k covariates matrix with non-null row and column names.}

\item{vec_idx}{numeric vector of sample/entry/pool indexes to extract from list_merged}

\item{verbose}{show genotype, phenotype, and covariate subsetting messages? (Default=FALSE)}
}
\value{
\itemize{
\item Ok:
\itemize{
\item $G: numeric n samples x p loci-alleles matrix of allele frequencies with non-null row and column names.
Row names can be any string of characters which identify the sample or entry or pool names.
Column names need to be tab-delimited, where first element refers to the chromosome or scaffold name,
the second should be numeric which refers to the position in the chromosome/scaffold, and
subsequent elements are optional which may refer to the allele identifier and other identifiers.
\item $list_pheno:
$y: named vector of numeric phenotype data
$pop: population or groupings corresponding to each element of y
$trait_name: name of the trait
\item $COVAR: numeric n samples x k covariates matrix with non-null row and column names.
}
\item Err: gpError
}
}
\description{
Subset the list of merged genotype and phenotype data using a vector of sample/entry/pool indexes.
}
\examples{
list_sim = fn_simulate_data(n_pop=3, verbose=TRUE)
G = fn_load_genotype(fname_geno=list_sim$fname_geno_vcf)
rownames(G)[1] = "entry_exclude_me"
rownames(G)[2] = "entry_exclude_me_too"
list_pheno = fn_load_phenotype(fname_pheno=list_sim$fname_pheno_tsv)
COVAR = matrix(stats::rnorm(n=(10*nrow(G))), nrow=nrow(G))
rownames(COVAR) = rownames(G); colnames(COVAR) = paste0("covariate_", 1:ncol(COVAR))
list_merged = fn_merge_genotype_and_phenotype(G=G, list_pheno=list_pheno, COVAR=COVAR, verbose=TRUE)
vec_idx = which(list_merged$list_pheno$pop == list_merged$list_pheno$pop[1])
list_merged_subset = fn_subset_merged_genotype_and_phenotype(list_merged=list_merged,
 vec_idx=vec_idx, verbose=TRUE)
}
