% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{foreach2}
\alias{foreach2}
\title{Enhanced version of foreach()}
\usage{
foreach2(
  .pass.wd = TRUE,
  .working.dir = NULL,
  .log.fn = NULL,
  .dump.fn = NULL,
  .parallel = TRUE,
  ...
)
}
\arguments{
\item{.pass.wd}{should the worker process set the same working directory as is on the master, before the job is executed? In most cases, \code{TRUE} will be the desired value (default).
Keep in mind that in some setups, e.g. with \code{library(doRedis)}, one might have the workers running on different machine.}

\item{.working.dir}{working directory to be set before the job is executed. Will override \code{.pass.wd} if set.}

\item{.log.fn}{if not \code{NULL}, standard and error output of the job will be saved into a log file with this file name.
Special \code{\%} sequences can be used, e.g. \code{"log-\%h_\%p.txt"}, see Details below.
Note that this parameter is evaluated in the context of the inner loop,
(!!! o to jsem se snazil ale to nefunguje bohuzel), so you might happily use the
foreach2() iterator variables in the expression!}

\item{.dump.fn}{if not \code{NULL}, debug dump will be saved upon an error, with this file name (the .rda extension will be added to it).
Special \code{\%} sequences can be used, e.g. \code{"dump-\%h_\%p"}, see Details below.
Note that this parameter is evaluated in the context of the inner loop (!!! o to jsem se snazil ale to nefunguje bohuzel),
so you might happily might use the \code{foreach2()} iterator variables in the expression!}

\item{...}{arguments passed to \code{\link[foreach]{foreach}()}.}
}
\description{
Enhanced version of \code{foreach()} that takes care of setting correct working directory, logging standard and error output to a log file, and dumping debug info upon error.
}
\details{
The arguments \code{log.fn} and \code{dump.fn} allow for special sequences:
\itemize{
\item \code{\%h} - hostname, i.e. the name of the machine where the worker job runs
\item \code{\%p} - process ID of the worker job
}
}
