% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{getwd.keepsym}
\alias{getwd.keepsym}
\title{Get working directory as user specified it (keeping symbolic links in the path)}
\usage{
getwd.keepsym()
}
\description{
Like getwd(), but keeps the path originally specified by the user, with symbolic links intact. This is a difference from \code{getwd()}, because on UNIX systems the path
might contain symlinks, and getwd() will then return a different path, with those symlinks resolved.
}
