% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{gh_quadrature}
\alias{gh_quadrature}
\title{Computes the 2D Gaussian quadrature rule of order 12}
\usage{
gh_quadrature(mu_n, sigma_n, mu_u, sigma_u, nodes2D, weights2D)
}
\arguments{
\item{mu_n}{mean of permanent shocks}

\item{mu_u}{mean of temp shocks}

\item{sigma_u}{sd of temp shocks}

\item{sigma}{sd of permanent shocks}
}
\description{
Computes the 2D Gaussian quadrature rule of order 12
}
