% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpData.R
\name{gpDataCheckReq}
\alias{gpDataCheckReq}
\title{internal function}
\usage{
gpDataCheckReq(gp, gpData)
}
\description{
check, whether gpData fulfills the requirements given by the gp object (gp$dataReq in particular),
i.e. whether all tables required by the formula are present, with correct grouping factors
}
