% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpData.R
\name{gpDataHasMainTable}
\alias{gpDataHasMainTable}
\title{Internal function to check whether gpData has a main table}
\usage{
gpDataHasMainTable(gpData)
}
\arguments{
\item{gpData}{object of class gpData}
}
\value{
TRUE/FALSE
}
\description{
Internal function to check whether gpData has a main table
}
