% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpData.R
\name{gpDataIsScaled}
\alias{gpDataIsScaled}
\title{Check whether gpData has been scaled already (gpDataPrepare() has been called on it)}
\usage{
gpDataIsScaled(gpData)
}
\arguments{
\item{gpData}{object of class gpData}
}
\value{
TRUE/FALSE
}
\description{
Check whether gpData has been scaled already (gpDataPrepare() has been called on it)
}
