% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpData.R
\name{gpDataPrepare}
\alias{gpDataPrepare}
\title{Prepares the object of class gpData for the model (scaling, conversion to matrices etc.)}
\usage{
gpDataPrepare(gp, gpData, scale.as = NULL)
}
\arguments{
\item{gp}{object of class \code{gp}}

\item{gpData}{object of class \code{gpData}; the data to be prepared (scaled etc.)}

\item{scale.as}{object of class \code{gpData}; if given, the data \code{gpData} will be scaled exactly the same way as the \code{scale.as} data.}
}
\description{
Prepares the object of class gpData for the model (scaling, conversion to matrices etc.)
}
\details{
Prepares the data for the model:
\itemize{
\item converts tables (data.frames) to matrices
\item scales the matrices (to mean = 0 and sd = 1) that need it (scale = TRUE in cov_funcs.R)
\itemize{
\item if the gp object already has scaled training data (gp$data), this function will take scaling from there, and
It issues an error, if there are any non-numeric columns etc.
}
}
}
