% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpData.R
\name{gpDataSubset}
\alias{gpDataSubset}
\title{Subset gpData along given factor and indices}
\usage{
gpDataSubset(x, fact = "1", ind)
}
\arguments{
\item{x}{object of class gpData. Does not matter if the data is scaled (gpDataPrepare()'d) or not,
it works on both.}

\item{fact}{character. Grouping factor name along which to index, or "1" for no grouping factor (the default),
which means indexing along the main table (or all tables, if there is no main table; since in the absence of main table
all tables must have one-to-one correspondence between their rows).}

\item{ind}{indices along given factor \code{fact} to subset from the dataset.  If \code{fact = "1"}, indices correspond to the
main table (or simply all tables, if no main table exists).}
}
\value{
gpData object, subsetted along the given factor and indices.
}
\description{
Subsets the gpData object along the given grouping factor \code{fact} (or along the main table if fact = "1"),
taking only the rows/levels with given \code{indices}.
}
