% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpData.R
\name{gpDetermineSize}
\alias{gpDetermineSize}
\title{Determine the size of the Gaussian Process}
\usage{
gpDetermineSize(gp)
}
\arguments{
\item{gp}{object of class gp}
}
\value{
A list with two elements:
\describe{
\item{size}{integer - size of the Gaussian Process (dimension of the covariance matrix; i.e. both number of rows and columns, since
the covariance matrix is a square matrix)}
\item{fact}{character - grouping factor corresponding to the dimension of the Gaussian process which defines the size; or "1" meaning "no grouping factor"}
}
}
\description{
Determine the size of the Gaussian Process, i.e. the dimension of the covariance matrix, as well as the corresponding grouping factor which defines this size
(or "1" if there is no such grouping factor).
}
\details{
\enumerate{
\item If all covariance components (except intercept) use the same grouping factor (true grouping factor, not "1"), then the dimension of the Gaussian Process is determined
by this factor, and the size is given by the number of levels of this factor.
\item In all other cases (no grouping factors, or some components with and some without a grouping factor, or multiple grouping factors used in different components),
the size is given by the number of rows of the main table (if the main table is not present, all tables necessarily have the same number of rows). The returned factor is "1",
meaning "no grouping factor".
}
}
