% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpFitCV.R
\name{gpGetCVModel}
\alias{gpGetCVModel}
\title{Get the cross-validation model for given fold}
\usage{
gpGetCVModel(gp, fold)
}
\arguments{
\item{gp}{GP model object}

\item{fold}{integer, number of the cross-validation fold}
}
\value{
GP model object for the fold \code{fold}
}
\description{
Get the cross-validation model for given fold
}
