% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpHyperpar.R
\name{gpHyperparCheck}
\alias{gpHyperparCheck}
\title{Check that hyperparameter vector \code{h} is within the limits \code{low} and \code{up}}
\usage{
gpHyperparCheck(gp, h, tol = sqrt(.Machine$double.eps), incl.fixed = FALSE)
}
\arguments{
\item{gp}{GP model object}

\item{h}{numeric vector to be checked}

\item{tol}{tolerance for numeric comparisons}

\item{incl.fixed}{does the vector \code{h} include the non-optimized (fixed) hyperparameters?}
}
\description{
Check that hyperparameter vector \code{h} is within the limits \code{low} and \code{up}
}
