% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpHyperpar.R
\name{gpHyperparCheckAll}
\alias{gpHyperparCheckAll}
\title{Check consistency of the hyperparameters}
\usage{
gpHyperparCheckAll(gp)
}
\arguments{
\item{gp}{GP model object}
}
\description{
Check consistency of the hyperparameters
}
