% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpHyperpar.R
\name{gpHyperparExportVector}
\alias{gpHyperparExportVector}
\title{export a column from hyperparameter table, for passing it to optimizer (thus skipping the fixed parameters)}
\usage{
gpHyperparExportVector(gp, col = "value")
}
\description{
export a column from hyperparameter table, for passing it to optimizer (thus skipping the fixed parameters)
}
