% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpHyperpar.R
\name{gpHyperparIdx}
\alias{gpHyperparIdx}
\title{Get the index of i-th optimized (non-fixed) hyperparameter (numbered from 1) in the table of hyperparameters}
\usage{
gpHyperparIdx(gp, i)
}
\arguments{
\item{gp}{GP model object}

\item{i}{index of the hyperparameter within non-fixed (optimized) hyperparameters}
}
\description{
Get the index of i-th optimized (non-fixed) hyperparameter (numbered from 1) in the table of hyperparameters
}
