% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpHyperpar.R
\name{gpHyperparImportVector}
\alias{gpHyperparImportVector}
\title{import hyperparameter vector \code{h}, given by the optimizer, into our hyperparameter table.}
\usage{
gpHyperparImportVector(gp, h, col = "value")
}
\description{
import hyperparameter vector \code{h}, given by the optimizer, into our hyperparameter table.
}
