% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpHyperpar.R
\name{gpHyperparList}
\alias{gpHyperparList}
\title{Converts a single column of the hyperparameter table to hierarchically structured list of lists of vectors (for each component and hyperparameter)}
\usage{
gpHyperparList(gp, col = "value")
}
\description{
Converts a single column of the hyperparameter table to hierarchically structured list of lists of vectors (for each component and hyperparameter)
}
