% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpHyperpar.R
\name{gpHyperparStartFromModel}
\alias{gpHyperparStartFromModel}
\title{Get the starting values of the hyperparameters from another model (gp0) wherever possible}
\usage{
gpHyperparStartFromModel(gp, gp0)
}
\arguments{
\item{gp}{GP model object to modify}

\item{gp0}{GP model to take the starting values from}
}
\value{
modified GP model object
}
\description{
Will set the hyperpar columns \code{start} and \code{value} based on \code{value} from model \code{gp0} wherever it will match the component, hyperparameter name and variable name.
The starting values that are not available in the model \code{gp0} will be kept as they were.
}
