% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpPack.R
\name{gpPack}
\alias{gpPack}
\title{Pack the model - reduce its size for memory or disk storage}
\usage{
gpPack(gp, maximum = FALSE)
}
\arguments{
\item{maximum}{should it be taken to maximum level? Default FALSE.}
}
\description{
Reduce model size by deleting auxiliary data structures which speed up the computations
but take a lot of space.
}
