% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{\%dopar2\%}
\alias{\%dopar2\%}
\title{Enhanced version of \verb{\%dopar\%}}
\usage{
obj \%dopar2\% ex
}
\description{
\if{html}{\out{<div class="sourceCode">}}\preformatted{Enhanced version of \code{\%dopar\%} that works with \code{foreach2()} and takes care of setting correct working directory, logging standard and error output to a log file, and dumping debug info upon error.
}\if{html}{\out{</div>}}
}
