% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmFit.R
\name{is_intercept_only_formula}
\alias{is_intercept_only_formula}
\title{Check for an intercept-only formula}
\usage{
is_intercept_only_formula(formula)
}
\arguments{
\item{formula}{A model formula.}
}
\value{
A logical scalar.
}
\description{
Returns \code{TRUE} when a formula has an intercept and no predictor terms,
e.g. \code{y ~ 1}.
}
\keyword{internal}
