% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculus.R
\name{is_numeric_finite}
\alias{is_numeric_finite}
\title{Tests if matrix M is numeric, with all finite, non-NA values. Works on dense as well as sparse matrices.}
\usage{
is_numeric_finite(M)
}
\description{
Tests if matrix M is numeric, with all finite, non-NA values. Works on dense as well as sparse matrices.
}
