% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/K_matrix.R
\name{K_matrix}
\alias{K_matrix}
\title{Construct covariance matrix (K) for given data and hyperparameters}
\usage{
K_matrix(
  gp,
  hyperpar = gpHyperparList(gp),
  x1 = gp$data,
  x2 = NULL,
  K.cache = NULL,
  components = NULL,
  comp_means = FALSE,
  comp_means_weights = NULL
)
}
\arguments{
\item{gp}{object of class \code{gp}.}

\item{hyperpar}{(optional) lists of hyperparameter values, as returned by \code{gpHyperparList()}; default is to use the hyperparameters from the \code{gp} object.}

\item{x1}{object of class \code{gpData} - data corresponding to rows of the resultant covariance matrix.
It is supposed to be correctly scaled already! (see \code{gpDataPrepare()})}

\item{x2}{(optional) object of class \code{gpData} - data corresponding to columns of the resultant covariance matrix.
If not supplied, \code{x1} is used and
the covariance matrix is square and symmetric.
It is supposed to be correctly scaled already! (see \code{gpDataPrepare()})}

\item{K.cache}{(optional) object returned by \code{K_cache()} function, to speed up repeated calls to \code{K_matrix()} and \code{dK_dhi()} functions.}

\item{components}{of the covariance formula. One of the following:
\itemize{
\item NULL : all components will be selected (the default)
\item character vector: components to be selected
\item formula: additive formula specifying components, same syntax as in \link[stats:update.formula]{update.formula}, e.g.
\tabular{ll}{
\code{~.}          \verb{   } \tab all components \cr
\code{~.-env-spat} \verb{   } \tab all components except \code{env} and \code{spat} \cr
\code{~spat+year}  \verb{   } \tab only components \code{spat} and \code{year} \cr
}
}}

\item{comp_means}{only meaningful for \code{x1} being the training dataset and \code{x2 = NULL}. If set to TRUE, then for each component, calculate vector of covariances
of the training data with the average of all the predictions from the training dataset, for that particular component. These vectors are
the rowmeans of training covariance matrices (that's why called component means). Save them in an attribute \code{comp_means} of K. These are later used
in the \code{predict} method.

\if{html}{\out{<div class="sourceCode">}}\preformatted{For comp_means = TRUE, x1 should be the training dataset. Perhaps it might make sense in other cases as well, but just make sure you know what you're doing :)
}\if{html}{\out{</div>}}}

\item{comp_means_weights}{\itemize{
\item don't have to sum up to 1, will be normalized
}}
}
\value{
Covariance matrix K:
\itemize{
\item Number of rows will be equal to \code{gpDataSize(x1, gp$GP_factor)}, where \code{gp$GP_factor} is the grouping factor corresponding to the dimension of the Gaussian process.
\item Analogically, if \code{x2} is supplied, the number of columns will be equal to \code{gpDataSize(x2, gp$GP_factor)}.
\item If \code{x2} is not supplied, the covariance matrix is square and symmetric, representing the covariances among all rows in \code{x1}.
}
}
\description{
This function is mostly used internally. In most common use, user will call gpFit() or predict(), which call this function themselves.
}
