% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{linterp}
\alias{linterp}
\title{Linear interpolation}
\usage{
linterp(x, z, c)
}
\arguments{
\item{z}{input parameter to linearly interpolate the x coordinates}

\item{c}{output vector to weight based on the input}
}
\description{
Linear interpolation
}
\details{
Returns a matrix of linearly interpolated values
}
