% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{loss}
\alias{loss}
\title{Solve model given parameters
gamma_0 intercept for retirement consumption
gamma_1 consumption in retirement as a linear function of
assets
beta discount factor
rho risk aversion}
\usage{
loss(gamma_0, gamma_1, beta, rho)
}
\description{
Solve model given parameters
gamma_0 intercept for retirement consumption
gamma_1 consumption in retirement as a linear function of
assets
beta discount factor
rho risk aversion
}
