% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculus.R
\name{mask}
\alias{mask}
\title{Mask dense matrix D to the structural non-zeros of sparse matrix S}
\usage{
mask(D, S)
}
\arguments{
\item{D}{dense matrix}

\item{S}{sparse matrix}
}
\value{
a sparse matrix, which is a matrix D but masked to the structural non-zeros of matrix S
}
\description{
Mask dense matrix D to the structural non-zeros of sparse matrix S
}
