% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mnfun.R
\name{mnfun}
\alias{mnfun}
\title{Compute mean of the Gaussian Process}
\usage{
mnfun(gp, data = gp$data, hyperpar = NULL)
}
\arguments{
\item{gp}{object of class \code{gp}.}

\item{data}{data along which to generate the gaussian process mean function.}

\item{hyperpar}{(optional) lists of hyperparameter values, as returned by \code{gpHyperparList()}; default is to use the hyperparameters from the \code{gp} object.}
}
\value{
Vector of the length of \code{gpDataSize(data, gp$GP_factor)}.
}
\description{
Compute mean of the Gaussian Process along given dataset. Currently just returns vector of zeros.
}
