% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{net_euler_diff}
\alias{net_euler_diff}
\title{Computes difference between expected marginal utility and
current marginal utility}
\usage{
net_euler_diff(
  c_now,
  x,
  c_next,
  const_scale_coh,
  const_add_coh,
  const_scale_consump,
  weights,
  R,
  p_noinc,
  beta,
  rho
)
}
\arguments{
\item{c_now}{current consumption}

\item{x}{current cash on hand}

\item{c_next}{proposed next consumption}

\item{const_scale_coh}{cash on hand scale param (see paper)}

\item{const_add_coh}{cash on hand addition param (see paper)}

\item{const_scale_consump}{consumption scale param (see paper)}

\item{weights}{Weights for weighted average from Gauss-Hermite
quadrature}

\item{R}{risk free rate}

\item{p_noinc}{probability of income equal to 0}

\item{beta}{discount factor}

\item{rho}{CRRA risk aversion parameter}
}
\description{
Computes difference between expected marginal utility and
current marginal utility
}
