% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred.R
\name{pred}
\alias{pred}
\title{Internal predict function}
\usage{
pred(
  gp,
  predx,
  same = FALSE,
  hyperpar = gpHyperparList(gp),
  components = NULL,
  comp_missing = c("avg", "none"),
  w = NULL,
  groupMeans = NULL,
  se.fit = TRUE,
  cov.fit = FALSE,
  maxn = Inf,
  Kx.cache = NULL,
  Kxx.cache = NULL,
  parallel = FALSE,
  log.fn = NULL,
  dump.fn = NULL,
  log.append = FALSE,
  tr.max.lines = 5,
  recursive = FALSE
)
}
\arguments{
\item{components}{character vector of names of the components of covariance that should be used for this prediction}

\item{se.fit}{return the standard errors of the predicted values?}

\item{cov.fit}{return the covariance matrix of the predicted values?}

\item{parallel}{use \code{parallel::parLapply()} on the registered default cluster for chunked predictions?}

\item{log.fn}{if not \code{NULL}, log parallel chunk jobs using \code{parallelJobWrapper()}}

\item{dump.fn}{if not \code{NULL}, dump debug info from failed parallel chunk jobs using \code{parallelJobWrapper()}}

\item{log.append}{logical; if \code{TRUE}, append to \code{log.fn} instead of overwriting it}

\item{tr.max.lines}{the \code{max.lines} parameter for \code{traceback}, i.e. the maximum number of lines printed per call when error occurs}
}
\value{
A matrix of predictions, always at the dimension of the Gaussian Process, with columns \code{f} and \code{f_SE} (if \code{se.fit = TRUE}).
If \code{cov.fit = TRUE}, returns a named list, \code{pred} will be the mentioned matrix and \code{cov} will be the full covariance matrix.
}
\description{
Internal predict function
}
