% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred_link.R
\name{predict_expand_link}
\alias{predict_expand_link}
\title{Expand the predictions by quantities derived using the inverse link function}
\usage{
predict_expand_link(pred, link = "logit", parname)
}
\arguments{
\item{pred}{data.frame or matrix of basic predictions to be expanded (\code{f}, etc.), as would be returned by predict.gp(type = "latent")}

\item{link}{character - name of link function to take inverse of - same as used by make.link()}

\item{parname}{character - name of the new parameter}
}
\value{
Expanded predictions.
}
\description{
Gets predictions of \code{f} from the Gaussian Process on the input, and from this computes also prediction for derived quantities,
in this case, using an inverse of given link function.
}
