% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likPois.R
\name{predPerfPois}
\alias{predPerfPois}
\title{Evaluation of predictive performance for poisson likelihood}
\usage{
predPerfPois(par, par.null)
}
\arguments{
\item{par}{the output of processing stage (stage 3) of the likelihood template. For the poisson likelihood in particular, it is
a list with two elements:
\describe{
\item{lambda}{numeric vector - predicted mean}
\item{y}{poisson response variable (counts)}
}}

\item{par.null}{the same as \code{par}, but for the null model (with intercept only). Only the \code{lambda} will be used from this list.}
}
\value{
list with the calculated metrics:
\itemize{
\item R2 - squared ordinary correlation between observed and predicted values. Scale from 0 to 1
\item R2_dev - pseudo R-squared based on poisson deviance. Scale from 0 to 1, 0 is null model (with intercept only)
\item R2_LR - McFadden's pseudo R-squared based on likelihood ratio. Scale from 0 to 1, 0 is null model (with intercept only)
\item R2_LRNCU - Nagelkerke, Cragg & Uhler's pseudo R-squared based on likelihood ratio, corrected for sample size. Scale from 0 to 1, 0 is null model (with intercept only)
}

Perfect prediction (the limit) has always value 1 in all these statistics.

Note that for models with more complicated likelihood the R2 might not make sense. The R2_* statistics are more general, and thus more robust sense-wise.
}
\description{
Evaluation of predictive performance for poisson likelihood
}
