% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{scale_cols}
\alias{scale_cols}
\title{scale() only selected columns}
\usage{
scale_cols(d, cols)
}
\arguments{
\item{d}{data.frame}

\item{cols}{character, column names}
}
\value{
the whole data.frame like \code{d}, with the columns \code{cols} scaled. The scaling attributes from scale() will
be saved as well.
}
\description{
scale() only selected columns
}
