% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{sim_given_params}
\alias{sim_given_params}
\title{Simulate consumption given parameters}
\usage{
sim_given_params(gamma_0, gamma_1, beta, rho)
}
\arguments{
\item{gamma_0}{intercept of retirement savings}

\item{gamma_1}{slope of portion of permanent income used in retirement savings}

\item{beta}{time discount factor}

\item{rho}{risk aversion for CRRA}
}
\description{
Simulate consumption given parameters
}
