% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{simulate_assets}
\alias{simulate_assets}
\title{Simulate N asset draws from log normal distribution}
\usage{
simulate_assets(N, mu, sigma)
}
\arguments{
\item{N}{number of draws}

\item{mu}{mean of distribution}

\item{sigma}{sd of distribution}
}
\description{
Simulate N asset draws from log normal distribution
}
