% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{simulate_income}
\alias{simulate_income}
\title{Simulate income process}
\usage{
simulate_income(N, T, P_init, G, sigma_n, sigma_u, p_noinc)
}
\arguments{
\item{N}{number of simulated agents}

\item{T}{number of time periods}

\item{G}{vector of growth rates (levels)}

\item{sigma_n}{standard deviation of permanent income shocks}

\item{sigma_u}{standard deviation of temporary income shocks}

\item{p_noinc}{probably income is equal to zero}
}
\description{
Simulate income process
}
