% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{simulate_lifecycle}
\alias{simulate_lifecycle}
\title{Simulate consumption / savings lifecycle problem}
\usage{
simulate_lifecycle(
  N,
  T,
  x_grid,
  P_init,
  G,
  sigma_n,
  sigma_u,
  gamma_0,
  gamma_1,
  mu_a,
  sigma_a,
  R,
  p_noinc,
  beta,
  rho
)
}
\arguments{
\item{N}{number of simulations}

\item{T}{number of time periods}

\item{G}{growth of permanent income}

\item{sigma_n}{sd of permanent income}

\item{sigma_u}{sd temporary income shocks}

\item{mu_a}{average log starting assets}

\item{sigma_a}{sd of log starting assets}

\item{R}{gross return on assets}

\item{p_noinc}{probability that income is zero}

\item{beta}{time discount factor}

\item{rho}{CRRA risk aversion}
}
\description{
Simulate consumption / savings lifecycle problem
}
