% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{solve_model}
\alias{solve_model}
\title{Solve the model given initial parameters}
\usage{
solve_model(init, ...)
}
\arguments{
\item{init}{initial vector of moments to use for the solution}

\item{...}{other parameters to pass to the optimizer}
}
\description{
Solve the model given initial parameters
}
\details{
Right now this assumes an identity weight matrix,
which is not what the authors use. They use optimal GMM, which
iterates back and forth on the weight matrix and the optimal
solution vector via a plugin estimator for the variance /
covariance matrix. Right now that is still to-do, but
adding a weight matrix to the loss function should be
a simple tweak.
}
