% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{startLog}
\alias{startLog}
\title{Redirect stdout and stderr to a log file}
\usage{
startLog(
  filename,
  tee = FALSE,
  all_or_nothing = TRUE,
  add_timestamp = FALSE,
  append = FALSE,
  autoWarn1 = TRUE
)
}
\arguments{
\item{filename}{character. Path to the log file.}

\item{tee}{logical. If \code{TRUE}, stream output to console as well (Unix only). currently on windows, STDERR cannot be split (i.e. will not be displayed on the console).}

\item{all_or_nothing}{logical. Makes only sense for tee = TRUE. When TRUE, perform "tee" only on systems where full funcionality is available}

\item{add_timestamp}{logical. Prepend timestamps using \code{ts} (Unix only, requires \code{ts} from moreutils).}

\item{append}{logical. If \code{TRUE}, append to \code{filename} instead of overwriting.}

\item{autoWarn1}{logical. If \code{TRUE}, it will set \code{options(warn = 1)} so that warnings are printed to the log immediately as well.}
}
\value{
A connection object used for logging, or \code{NULL} on failure.
}
\description{
Redirects R's standard output and messages (stderr)
to the specified file. On Unix systems, can optionally keep output visible on the
console using \code{tee} and add timestamps via the \code{ts} command.
}
\examples{
\dontrun{
LOG <- startLog("mylog.txt", tee = interactive(), append = TRUE)
# ... do work ...
endLog(LOG)
}

}
