% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{tryNull}
\alias{tryNull}
\title{A version of try() that returns NULL upon failure}
\usage{
tryNull(expr)
}
\arguments{
\item{expr}{An expression to evaluate.}
}
\value{
The value of \code{expr}, or \code{NULL} if evaluation fails with an error.
}
\description{
Evaluate an expression while trapping all errors. If an error occurs, it is
printed and \code{NULL} is returned. If the expression succeeds, its value is
returned unchanged.
}
\examples{
tryNull({
  12
  stop("blabla")
})

tryNull(log(10))

}
