% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpComponents.R
\name{validate_components}
\alias{validate_components}
\title{Validates the components argument - the selection of components}
\usage{
validate_components(gp, components)
}
\arguments{
\item{gp}{object of class gp}

\item{components}{one of the following:
\itemize{
\item character vector: components to be selected
\item NULL : all components will be selected
\item formula: additive formula specifying components, same syntax as in \link[stats:update.formula]{update.formula}, e.g.
\tabular{ll}{
\code{~.}          \tab all components \cr
\code{~.-env-spat} \tab all components except \code{env} and \code{spat} \cr
\code{~spat+year}  \tab only components \code{spat} and \code{year} \cr
}
}}
}
\value{
character vector of component names
}
\description{
Validates the components argument - the selection of components
}
