% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{consume}
\alias{consume}
\title{Consume out of cash on hand}
\usage{
consume(x, x_grid, cr, t)
}
\arguments{
\item{x}{cash on hand per unit of permanent income}

\item{x_grid}{grid we are using for linear interpolation}

\item{cr}{optimal consumption rule}

\item{t}{integer indicating time period to reference for consumption
solved for with consumption_rule function}
}
\description{
Consume out of cash on hand
}
\details{
This computes consumption given income
assets and return on assets by first computing cash
on hand and then passing to the optimal consumption
rule.
}
