% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv_eval.R
\name{cv_eval_bern}
\alias{cv_eval_bern}
\title{Evaluation of cross-validation statistics for bernoulli (0/1) response variable and predicted probability}
\usage{
cv_eval_bern(p, y)
}
\arguments{
\item{p}{numeric vector - predicted probability}

\item{y}{bernoulli response variable (vector of 0s and 1s)}
}
\value{
list with the calculated metrics
}
\description{
Calculates AUC, TSS, and R2_dev (binomial deviance)
}
