% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{fn_G_split_off_alternative_allele}
\alias{fn_G_split_off_alternative_allele}
\title{Split trailing alternative allele from a numeric (allele frequency) matrix if it exists}
\usage{
fn_G_split_off_alternative_allele(G, verbose = FALSE)
}
\arguments{
\item{G}{n samples x p loci matrix of allele frequencies (numeric ranging from 0 to 1) with non-null row and column names.
Row names can be any string of characters which identify the sample or entry or pool names.
Column names need to be tab-delimited, where first element refers to the chromosome or scaffold name,
the second should be numeric which refers to the position in the chromosome/scaffold, and
subsequent elements are optional which may refer to the allele identifier and other identifiers.}

\item{verbose}{show genotype data alternative allele splitting messages? (Default=FALSE)}
}
\value{
Ok:
$G: n samples x p loci matrix of allele frequencies excluding trailing allele (numeric ranging from 0 to 1)
$G_alt: n samples x p loci matrix of allele frequencies of the trailing allele (numeric ranging from 0 to 1)
Err: grError
}
\description{
Split trailing alternative allele from a numeric (allele frequency) matrix if it exists
}
\examples{
G_ref = simquantgen::fn_simulate_genotypes(verbose=TRUE)
G_alt = 1 - G_ref; colnames(G_alt) = gsub("allele_1$", "allele_2", colnames(G_alt))
G_refalt = cbind(G_ref, G_alt)
list_G_G_alt = fn_G_split_off_alternative_allele(G=G_refalt, verbose=TRUE)
}
