% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{fn_simulate_data}
\alias{fn_simulate_data}
\title{Simple wrapper of simquantgen simulation of 10 standard normally distributed QTL additive effects
on 5-chromosome genome and a single trait at 50\% heritability}
\usage{
fn_simulate_data(
  n = 100,
  l = 1000,
  ploidy = 2,
  n_alleles = 2,
  min_depth = 5,
  max_depth = 1000,
  n_pop = 1,
  n_effects = 10,
  h2 = 0.5,
  seed = 12345,
  save_pheno_tsv = TRUE,
  save_geno_vcf = TRUE,
  save_geno_tsv = FALSE,
  save_geno_rds = FALSE,
  non_numeric_Rds = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{n}{number of samples (Default=100)}

\item{l}{number of loci (Default=1000)}

\item{ploidy}{ploidy level which can refer to the number of haploid genomes to simulate pools (Default=2)}

\item{n_alleles}{maximum number of alleles per locus (Default=2)}

\item{min_depth}{minimum depth per locus (Default=5)}

\item{max_depth}{maximum depth per locus (Default=5000)}

\item{n_pop}{number of randomly assigned population groupings (Default=1)}

\item{n_effects}{number of additive QTL effects to simulate (Default=10)}

\item{h2}{narrow-sense heritability to simulate (Default=0.5)}

\item{seed}{randomisation seed for replicability (Default=12345)}

\item{save_pheno_tsv}{save the phenotype data as a tab-delimited file? (Default=TRUE)}

\item{save_geno_vcf}{save the genotype data as a vcf file? (Default=TRUE)}

\item{save_geno_tsv}{save the genotype data as a tab-delimited allele frequency table file? (Default=FALSE)}

\item{save_geno_rds}{save the named genotype matrix as an Rds file? (Default=FALSE)}

\item{non_numeric_Rds}{save non-numeric Rds genotype file? (Default=FALSE)}

\item{verbose}{show genotype and phenotype data simulation messages? (Default=FALSE)}
}
\value{
Ok:
$fname_geno_vcf: filename of the simulated genotype data as a vcf file
$fname_geno_tsv: filename of the simulated genotype data as a tab-delimited allele frequency table file
$fname_geno_rds: filename of the simulated named genotype matrix as an Rds file
$fname_pheno_tsv: filename of the simulated phenotype data as a tab-delimited file
Err: gpError
}
\description{
Simple wrapper of simquantgen simulation of 10 standard normally distributed QTL additive effects
on 5-chromosome genome and a single trait at 50\% heritability
}
\examples{
list_sim = fn_simulate_data(verbose=TRUE, 
 save_geno_vcf=TRUE, save_geno_tsv=TRUE, save_geno_rds=TRUE, save_pheno_tsv=TRUE)
}
