% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{gp}
\alias{gp}
\title{Main genomic prediction cross-validation within and across populations}
\usage{
gp(args)
}
\arguments{
\item{args}{list of arguments or inputs
$fname_geno: filename of the input genotype file (see ?fn_load_genotype for details)
$fname_pheno: filename of the input phenotype file (see ?fn_load_phenotype for details)
$population: names of the population to used in within population k-fold cross-validation
$fname_covar: filename of the covariate data (TODO: implement fn_load_covariate)
$dir_output: output directory into which all temporary and final output files will be written
$geno_fname_snp_list: filename of the SNP list for filtering the genotype data
(see ?fn_filter_genotype for details)
$geno_ploidy: expected ploidy level of the genotype file (see ?fn_load_genotype for details)
$geno_bool_force_biallelic: force all loci including multi-allelic ones to be biallelic?
(see ?fn_load_genotype for details)
$geno_bool_retain_minus_one_alleles_per_locus: remove the trailing allele per locus?
(see ?fn_load_genotype for details)
$geno_min_depth: minimum depth per allele (see ?fn_load_genotype for details)
$geno_max_depth: maximum depth per allele (see ?fn_load_genotype for details)
$geno_maf: minimum allele frequency (see ?fn_filter_genotype for details)
$geno_sdev_min: minimum allele frequency standard deviation (see ?fn_filter_genotype for details)
$geno_max_n_alleles: maximum number of alleles per locus (see ?fn_filter_genotype for details)
$geno_max_sparsity_per_locus: maximum sparsity or fraction of missing data per locus
(see ?fn_filter_genotype for details)
$geno_frac_topmost_sparse_loci_to_remove: fraction of the total number of loci from which the
top most sparse loci will be removed (see ?fn_filter_genotype for details)
$geno_n_topmost_sparse_loci_to_remove: number of top most sparse loci to be removed
(see ?fn_filter_genotype for details)
$geno_max_sparsity_per_sample:  maximum sparsity or fraction of missing data per sample
(see ?fn_filter_genotype for details)
$geno_frac_topmost_sparse_samples_to_remove: fraction of the total number of samples from which
the top most sparse samples will be removed (see ?fn_filter_genotype for details)
$geno_n_topmost_sparse_samples_to_remove: number of top most sparse samples to be removed
(see ?fn_filter_genotype for details)
$pheno_sep: delimiter used in the phenotype file (see ?fn_load_phenotype for details)
$pheno_header: does the phenotype file have a header line? (see ?fn_load_phenotype for details)
$pheno_idx_col_id: column number in the phenotype file corresponding to the sample names
(see ?fn_load_phenotype for details)
$pheno_idx_col_pop: column number in the phenotype file corresponding to the population/grouping names
(see ?fn_load_phenotype for details)
$pheno_idx_col_y: column number in the phenotype file corresponding to the numeric phenotype data
(see ?fn_load_phenotype for details)
$pheno_na_strings: strings of characters corresponding to missing data in the phenotype file
(see ?fn_load_phenotype for details)
$pheno_bool_remove_NA: remove samples missing phenotype data in the phenotype file?
(see ?fn_load_phenotype for details)
$bool_within: perform within population k-fold cross-validation?
(see ?fn_cross_validation_within_population for details)
$bool_across: perform across populations cross-validations?
(see ?fn_cross_validation_across_populations_bulk,
?fn_cross_validation_across_populations_bulk,
?fn_cross_validation_across_populations_pairwise, and
?fn_cross_validation_across_populations_lopo for details)
$n_folds: number of folds or training and validation sets for within population k-fold cross validation
(see ?fn_cross_validation_within_population for details)
$n_reps: number of replications of random shuffling samples for each k-fold cross-validation
in within population cross-validation (see ?fn_cross_validation_within_population for details)
$vec_models_to_test:
$bool_parallel:
$max_mem_Gb:
$n_threads:
$verbose: show messages?}
}
\value{
Ok:
$TRAIT_NAME: name of the phenotype or trait used to fit single-trait genomic prediction models (multi-trait models will be implemented in genomic_breeding)
$POPULATION: name of the population used in within population genomic prediction k-fold cross-validation
$METRICS_WITHIN_POP (row-binded df_metrics across populations, reps, folds, and models):
$rep: replication number
$fold: fold number
$model: genomic prediction model name
$pop_training: population/s used in the training set (separated by commas if more than 1)
$pop_validation: population/s used in the validation set (separated by commas if more than 1)
$duration_mins: time taken in minutes to fit the genomic prediction model and assess the prediction accuracies
$n_non_zero: number of non-zero estimated effects (effects greater than machine epsilon ~2.2e-16)
$mbe: mean bias error
$mae: mean absolute error
$rmse: root mean squared error
$r2: coefficient of determination
$corr: Pearson's product moment correlation
$power_t10: fraction of observed top 10 phenotype values correctly predicted
$power_b10: fraction of observed bottom 10 phenotype values correctly predicted
$var_pred: variance of predicted phenotype values (estimator of additive genetic variance)
$var_true: variance of observed phenotype values (estimator of total phenotypic variance)
$h2: narrow-sense heritability estimate
$YPRED_WITHIN_POP (row-binded df_y_validation across populations, reps, folds, and models):
$rep: replication number
$fold: fold number
$model: genomic prediction model name
$pop_training: population/s used in the training set (separated by commas if more than 1)
$id: names of the samples/entries/pools,
$pop_validation: population from which the sample/entry/pool belongs to
$y_true: observed phenotype values
$y_pred: predicted phenotype values
$METRICS_ACROSS_POP_BULK:
$rep: replication number
$fold: fold number
$model: genomic prediction model name
$pop_training: population/s used in the training set (separated by commas if more than 1)
$pop_validation: population/s used in the validation set (separated by commas if more than 1)
$duration_mins: time taken in minutes to fit the genomic prediction model and assess the prediction accuracies
$n_non_zero: number of non-zero estimated effects (effects greater than machine epsilon ~2.2e-16)
$mbe: mean bias error
$mae: mean absolute error
$rmse: root mean squared error
$r2: coefficient of determination
$corr: Pearson's product moment correlation
$power_t10: fraction of observed top 10 phenotype values correctly predicted
$power_b10: fraction of observed bottom 10 phenotype values correctly predicted
$var_pred: variance of predicted phenotype values (estimator of additive genetic variance)
$var_true: variance of observed phenotype values (estimator of total phenotypic variance)
$h2: narrow-sense heritability estimate
$YPRED_ACROSS_POP_BULK:
$rep: replication number
$fold: fold number
$model: genomic prediction model name
$pop_training: population/s used in the training set (separated by commas if more than 1)
$id: names of the samples/entries/pools,
$pop_validation: population from which the sample/entry/pool belongs to
$y_true: observed phenotype values
$y_pred: predicted phenotype values
$METRICS_ACROSS_POP_PAIRWISE:
$rep: replication number
$fold: fold number
$model: genomic prediction model name
$pop_training: population/s used in the training set (separated by commas if more than 1)
$pop_validation: population/s used in the validation set (separated by commas if more than 1)
$duration_mins: time taken in minutes to fit the genomic prediction model and assess the prediction accuracies
$n_non_zero: number of non-zero estimated effects (effects greater than machine epsilon ~2.2e-16)
$mbe: mean bias error
$mae: mean absolute error
$rmse: root mean squared error
$r2: coefficient of determination
$corr: Pearson's product moment correlation
$power_t10: fraction of observed top 10 phenotype values correctly predicted
$power_b10: fraction of observed bottom 10 phenotype values correctly predicted
$var_pred: variance of predicted phenotype values (estimator of additive genetic variance)
$var_true: variance of observed phenotype values (estimator of total phenotypic variance)
$h2: narrow-sense heritability estimate
$YPRED_ACROSS_POP_PAIRWISE:
$rep: replication number
$fold: fold number
$model: genomic prediction model name
$pop_training: population/s used in the training set (separated by commas if more than 1)
$id: names of the samples/entries/pools,
$pop_validation: population from which the sample/entry/pool belongs to
$y_true: observed phenotype values
$y_pred: predicted phenotype values
$METRICS_ACROSS_POP_LOPO:
$rep: replication number
$fold: fold number
$model: genomic prediction model name
$pop_training: population/s used in the training set (separated by commas if more than 1)
$pop_validation: population/s used in the validation set (separated by commas if more than 1)
$duration_mins: time taken in minutes to fit the genomic prediction model and assess the prediction accuracies
$n_non_zero: number of non-zero estimated effects (effects greater than machine epsilon ~2.2e-16)
$mbe: mean bias error
$mae: mean absolute error
$rmse: root mean squared error
$r2: coefficient of determination
$corr: Pearson's product moment correlation
$power_t10: fraction of observed top 10 phenotype values correctly predicted
$power_b10: fraction of observed bottom 10 phenotype values correctly predicted
$var_pred: variance of predicted phenotype values (estimator of additive genetic variance)
$var_true: variance of observed phenotype values (estimator of total phenotypic variance)
$h2: narrow-sense heritability estimate
$YPRED_ACROSS_POP_LOPO:
$rep: replication number
$fold: fold number
$model: genomic prediction model name
$pop_training: population/s used in the training set (separated by commas if more than 1)
$id: names of the samples/entries/pools,
$pop_validation: population from which the sample/entry/pool belongs to
$y_true: observed phenotype values
$y_pred: predicted phenotype values
$GENOMIC_PREDICTIONS
$id: names of the samples/entries/pools which has genotype data but missing or outlying phenotype data
$pop: population or grouping the samples/entries/pools belong to
$y_true: missing observed phenotype data
$y_pred: predicted phenotype data using the best performing genomic prediction model
$model: best performing model based on within population cross-validation used to predict the missing phenotype data
$ADDITIVE_GENETIC_EFFECTS
$b: named numeric vector of SNP effects (or genotype effects for gBLUP),
where the names consist of tab-delimited chromosome/scaffold, position and if present, alleles.
$model: best performing genomic prediction model based on within population cross-validation
which yielded these effects.
Err: gpError
}
\description{
Main genomic prediction cross-validation within and across populations
}
