% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gp.R
\name{gp}
\alias{gp}
\title{Construct gaussian process model}
\usage{
gp(
  f,
  data,
  lik,
  lik.hyperpar = NULL,
  lik.formula = NULL,
  lik.reindex2main = TRUE,
  W.type = c("diag", "bdiag", "other"),
  predictor.fun = NULL,
  response.parname = NULL,
  link = NULL
)
}
\arguments{
\item{f}{formula for the covariance matrix}

\item{data}{object of class gpData, created by the gpData() function}

\item{lik}{an object of class likTemp, representing the user-defined template of the likelihood function \code{p(y|f)}.

ORIG DOC TO BE DELETED - MAYBE USE PART OF THESE FOR THE DOCUMENTATION OF CLASS likTemp
either a single function, or a named list of three functions (three-stage template); these are two ways to supply
the user-defined negative log likelihood function p(y|f):
\enumerate{
\item \code{negLogLik} is a single function \code{f(data, par)}, taking two parameters:
\describe{
\item{data}{object of class \code{gpData}, containing both predictors and a response variable}
\item{\code{par}}{a list of parameters (numeric scalars or vectors).
Always includes \code{f}, a numeric vector - result of the gaussian process. The \code{par$f} vector can be either of the dimension of the gaussian process,
or the main table - see the \code{negLogLik.reindex2main} parameter. The \code{par} parameters list can optionally include other parameters, see the \code{negLogLik.hyperpar} argument below.
}
}
\item Logic is similar as in the previous case, but the \code{negLogLik} is supplied as a chain of three consecutive functions (three-stage template). Then, \code{negLogLik} is a named list of three functions:
\describe{
\item{\code{pred.link}}{either character or \code{function(data, par)}. This is predictor & link component. Processes the parameters into other parameters;
is also used by \code{predict(type = "response")} for predictions,
therefore, the \code{data} might not contain the response variable (thus, the response variable must not be used by this function). Should be either a character string representing
the name of a link function    (this will be the most frequent case; see !!! for supported link functions), or a \code{function(data, par)} (see above for parameter description !!!).
This function will return a named list of parameters \code{par} that will be passed on to the next stage.}
\item{\code{process}}{\code{function(data, par)}, this function processes the parameters and the \code{data} (which always contain the response variable here) and shapes both
parameters and data for particular likelihood function. This function should return a named list required by particular likelihood function (see !!!).
This component is used by cross-validation and Laplace-Fisher approximation.}
\item{\code{lik}}{function evaluating the likelihood. Either character representing name of likelihood function (see !!! for supported values), or custom \code{function(x)} taking \code{x} from
the previous stage.}
}
}}

\item{lik.hyperpar}{optional, a named list of extra likelihood hyperparameters for the \code{negLogLik} function (named list of numeric scalars or vectors); this doesn't include the special
parameter \code{f} (the Gaussian process). This effectively defines the likelihood hyperparameters, named ".lik" in the hyperparameter table.}

\item{lik.formula}{optional, formula related to the supplied \code{negLogLik} function. This is not used in any way, except it is stored in the object and printed in the summary,
so it has only informational purpose.}

\item{lik.reindex2main}{should the \code{par$f} parameter of the \code{negLogLik} function be reindexed to the rows of the main table?
If \code{FALSE}, the \code{par$f} parameter will be kept at the dimension of the Gaussian Process (as reported by \code{gpDetermineSize()}).
If the Gaussian process is running at the dimension of the main table, then it does not matter.}

\item{W.type}{the type of the hessian matrix (\code{W}) of the negative log likelihood function (as defined by the \code{lik} parameter)
\describe{
\item{\code{"diag"}}{(default) - the hessian matrix (\code{W}) is diagonal. Most common case, fastest computation. This is usually the case when the user-defined likelihood factorizes over individual \eqn{f_i}'s.}
\item{\code{"bdiag"}}{the hessian matrix (\code{W}) is block-diagonal. This is usually the case when the user-defined likelihood factorizes over mutually disjoint sets of \eqn{f_i}'s.}
\item{\code{"other"}}{the hessian matrix (\code{W}) is a general matrix.}
}
One should keep the setting as specific as possible to keep the computation efficient.}
}
\value{
object of class gp - the gaussian process model. Some important slots:
\describe{
\item{$hyperpar}{hyperparameter table (see XXX). User is invited to tweak these settings.}
\item{$obsdata}{observed training data, in the original scale, as user supplied them}
\item{$data}{training data prepared for model fit - converted to matrices, and scaled (standardized) where appropriate}
}
}
\description{
Construct gaussian process model
}
