% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gp.R
\name{gp}
\alias{gp}
\title{Construct gaussian process model}
\usage{
gp(f, data, logLik)
}
\arguments{
\item{f}{formula for the covariance matrix}

\item{data}{object of class gpData, created by the gpData() function}

\item{logLik}{blikelihood function taking two parameters: \code{data} (object of class gpData), and \code{f}, a numeric vector - result of the gaussian process}
}
\value{
object of class gp - the gaussian process model. Some important slots:
\describe{
\item{$hyperpar}{hyperparameter table (see XXX). User is invited to tweak these settings.}
\item{$obsdata}{observed training data, in the original scale, as user supplied them}
\item{$data}{training data prepared for model fit - converted to matrices, and scaled (standardized) where appropriate}
}
}
\description{
Construct gaussian process model
}
