% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gp.R
\name{gp}
\alias{gp}
\title{Construct gaussian process model}
\usage{
gp(f, data, logLik, logLik.reindex2main = TRUE)
}
\arguments{
\item{f}{formula for the covariance matrix}

\item{data}{object of class gpData, created by the gpData() function}

\item{logLik}{likelihood function taking two parameters: \code{data} (object of class gpData), and \code{f}, a numeric vector - result of the gaussian process.
The \code{f} vector can be either of the dimension of the gaussian process, or the main table - see the \code{logLik.reindex2main} parameter.}

\item{logLik.reindex2main}{should the \code{f} parameter of the \code{logLik} function be reindexed to the rows of the main table?
If \code{FALSE}, the \code{f} parameter will be kept at the dimension of the Gaussian Process (as reported by \code{gpDetermineSize()}).
If the Gaussian process is running at the dimension of the main table, then it does not matter.}
}
\value{
object of class gp - the gaussian process model. Some important slots:
\describe{
\item{$hyperpar}{hyperparameter table (see XXX). User is invited to tweak these settings.}
\item{$obsdata}{observed training data, in the original scale, as user supplied them}
\item{$data}{training data prepared for model fit - converted to matrices, and scaled (standardized) where appropriate}
}
}
\description{
Construct gaussian process model
}
