% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpFit.R
\name{gpFit}
\alias{gpFit}
\title{Fit a Gaussian process model}
\usage{
gpFit(
  gp,
  method = c("Laplace", "Laplace-Fisher"),
  opt.h = TRUE,
  two.stage = TRUE,
  stages = 1:2,
  stage1low = 0.01,
  h = NULL,
  use.prior = TRUE,
  hessian = FALSE,
  opt.method = "L-BFGS-B",
  opt.control = list(factr = 1e+10, maxit = 190),
  fisher.options = list(sampling = FALSE, samples = 1000),
  verbose = FALSE,
  use_f_start = TRUE,
  weights = NULL,
  grad.computation = opt.h,
  warn.exclude = NULL,
  recursive = FALSE,
  ...
)
}
\arguments{
\item{gp}{GP model object.}

\item{method}{character; fitting method. Either \code{"Laplace"} (Laplace approximation), or \code{"Laplace-Fisher"} (Laplace-Fisher approximation).
See \code{fisher.options} for the latter method.
Laplace-Fisher approximation is the solution for when the hessian of the negative log likelihood is not positive semi-definite.}

\item{opt.h}{logical; should hyperparameters be optimised? Default TRUE; Only rarely you may want to say FALSE.}

\item{two.stage}{logical; if TRUE (the default), the two-stage optimization method will be applied (see Details below).}

\item{stages}{numerical vector - the optimization stages to be performed. Only relevant for \code{two.stage = TRUE}. Default is \code{1:2}, i.e. both stages.
See details below.

If you are running only stage 2 separately, which would make sense only in rare cases (perhaps \code{two.stage = FALSE} is what you need in most of those cases),
note that stage 2 doesn't start from \code{gp$hyperpar$start}, but from \code{gp$hyperpar$value}.}

\item{stage1low}{the low threshold for the sigma2 parameter of the components for which the two-stage fitting applies.}

\item{h}{optional; a numeric vector of starting values of hyperparameters (untransformed, i.e. on their most natural scale). If \code{NULL},
the starting values are taken from \code{gp} hyperparameter table - from the \code{gp$hyperpar[,"start"]} column when \code{opt.h = TRUE}, or from
the \code{gp$hyperpar[,"value"]} column when \code{opt.h = FALSE}.}

\item{use.prior}{logical; if TRUE, likelihood + prior (posterior) is optimized; if FALSE, only likelihood is optimized.}

\item{hessian}{logical; compute and return the Hessian during the hyperparameter optimization. Warning: may be heavy on CPU time! Default FALSE.}

\item{opt.method}{the method used by \code{optim(method = )}}

\item{opt.control}{list of control parameters passed to the \code{optim(control = )}}

\item{fisher.options}{list of the following options specifying how to generate the Fisher information matrix (applies when \code{method = "Laplace-Fisher"}):
\describe{
\item{\code{sampling}}{should sampling be used to create the Fisher information matrix? If \code{FALSE} (the default), likelihood-specific method must be implemented.
Sampling can be used when no such method is implemented (but it is slower).}
\item{\code{samples}}{how many samples should be generated for the sampling method?}
}}

\item{verbose}{logical}

\item{use_f_start}{logical; internal optimization - start each Laplace approximation from the optimized \code{f} vector from the previous hyperparameter iteration.}

\item{weights}{numeric or \code{NULL}; observation weights (not implemented).}

\item{grad.computation}{logical; compute gradients during the gaussian process optimization (required when
\code{opt.h = TRUE}).}

\item{warn.exclude}{character vector; vector of warnings that should be disabled. Allowed values are: \code{"warn.stage2.start"}. Default \code{NULL}.}

\item{recursive}{internal option; do not use.}

\item{...}{further arguments to be passed to fitting method (e.g. \code{gpFitLaplace()})}
}
\value{
Returns the input \code{gp} object with the model fit (stored in \code{gp$fit}) and updated hyperparameters.

\emph{NOTE!!} The returned object is usually large, because it contains pre-calculated matrices that are used for predictions. If you need to save the model,
call \code{gpPack()} first to remove these; when you need to use the model again, call \code{gpUnpack()} to restore these matrices.
}
\description{
Fit a Gaussian process model
}
\details{
The two-stage fitting method (\code{two.stage = TRUE}) is a special optimization procedure (Telenský et al 2026) where the optimization is split into two stages:
\enumerate{
\item In the first stage, sigma2 parameters are not allowed to go too close to zero (as per the \code{stage1low} parameter). This way, the other hyper-parameters in the same component have time
to optimize themselves before the sigma2 parameter could shut the whole component down to zero. (This is only applied to components
that have also other hyper-parameters apart from sigma2.)
\item In the second stage, the restriction applied in the first stage is lifted and all the sigma2 hyper-parameters are allowed to go further down to zero.
}

In some cases, where user tries to optimize the CPU time by manually splitting the procedure in multiple steps, it might be a good idea to run both stages only in the last step,
and use only stage 1 in all the previous steps (see Vignette XXX).

If you are running only stage 2 separately, which would make sense only in rare cases (perhaps \code{two.stage = FALSE} is what you need in most of those cases),
note that stage 2 doesn't start from \code{gp$hyperpar$start}, but from \code{gp$hyperpar$value}.
}
\examples{
\dontrun{
gp <- gp(...) # construct gp object
gp <- gpFit(gp)
}

}
