% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpFit.R
\name{gpFit}
\alias{gpFit}
\title{Fit a Gaussian process model}
\usage{
gpFit(
  gp,
  h = NULL,
  opt.h = TRUE,
  two.stage = TRUE,
  stages = 1:2,
  stage1low = 0.01,
  use.prior = TRUE,
  hessian = FALSE,
  opt.control = list(factr = 1e+10, maxit = 190),
  verbose = FALSE,
  method = c("Laplace"),
  use_f_start = TRUE,
  weights = NULL,
  grad.computation = TRUE,
  recursive = FALSE
)
}
\arguments{
\item{gp}{GP model object.}

\item{h}{optional; a numeric vector of starting values of hyperparameters (untransformed, i.e. on their most natural scale). If \code{NULL},
the starting values are taken from \code{gp} hyperparameter table (\code{gp$hyperpar[,"start"]}).}

\item{opt.h}{logical; should hyperparameters be optimised? Default TRUE; Only rarely you may want to say FALSE.}

\item{two.stage}{logical; if TRUE, the two-stage optimization method will be applied (see Details below).}

\item{stages}{numerical vector - the optimization stages to be performed. Only valid for \code{two.stage = TRUE}. Default is \code{1:2}, i.e. both stages.}

\item{stage1low}{the low threshold for the sigma2 parameter of the components for which the two-stage fitting applies.}

\item{use.prior}{logical; if TRUE, likelihood + prior (posterior) is optimized; if FALSE, only likelihood is optimized.}

\item{hessian}{logical; compute and return the Hessian during the optimization. Warning: may be heavy on CPU time! Default FALSE.}

\item{opt.control}{list of control parameters passed to the \code{optim(control = )}}

\item{verbose}{logical}

\item{method}{character; fitting method. Currently \code{"Laplace"} supported (Laplace approximation).}

\item{use_f_start}{logical; internal optimization - start each Laplace approximation from the optimized \code{f} vector from the previous hyperparameter iteration.}

\item{weights}{numeric or \code{NULL}; observation weights (not implemented).}

\item{grad.computation}{logical; compute gradients during the gaussian process optimization (required when
\code{opt.h = TRUE}).}

\item{recursive}{internal option; do not use.}
}
\value{
Returns the input \code{gp} object with the model fit (stored in \code{gp$fit}) and updated hyperparameters.

\emph{NOTE!!} The returned object is usually large, because it contains pre-calculated matrices that are used for predictions. If you need to save the model,
call \code{gpPack()} first to remove these; when you need to use the model again, call \code{gpUnpack()} to restore these matrices.
}
\description{
Fit a Gaussian process model
}
\details{
The two-stage fitting method (\code{two.stage = TRUE}) is a special optimization procedure (Telenský et al 2026) where the optimization is split into two stages:
\enumerate{
\item In the first stage, sigma2 parameters are not allowed to go too close to zero (as per the \code{stage1low} parameter). This way, the other hyper-parameters in the same component have time
to optimize themselves before the sigma2 parameter could shut the whole component down to zero. (This is only applied to components
that have also other hyper-parameters apart from sigma2.)
\item In the second stage, the restriction applied in the first stage is lifted and all the sigma2 hyper-parameters are allowed to go further down to zero.
}

In some cases, where user tries to optimize the CPU time by manually splitting the procedure in multiple steps, it might be a good idea to run the stage 2 only in the last step,
and use only stage 1 in all the previous steps (see Vignette XXX).
}
\examples{
\dontrun{
gp <- gp(...) # construct gp object
gp <- gpFit(gp)
}

}
