% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpFitCV.R
\name{gpFitCV}
\alias{gpFitCV}
\title{Perform an N-fold cross-validation of a Gaussian process model}
\usage{
gpFitCV(
  gp,
  fold.col,
  fold.fact = "1",
  folds = NULL,
  start.from.model = NULL,
  parallel = TRUE,
  fn.prefix = "",
  log.fn = if (parallel) "log-fold\%f-\%h-\%p.txt" else NULL,
  dump.fn = if (parallel) "dump-fold\%f-\%h_\%p" else NULL,
  tr.max.lines = 5,
  ...
)
}
\arguments{
\item{gp}{GP model object}

\item{fold.col}{either a name of a column in the main table, or a vector along factor \code{fold.fact}.
The column (or the supplied vector) must be a vector of integers from \code{1} to \code{N} (\code{N} being the number of folds),
specifying the number of the cross-validation fold the given record belongs to.}

\item{fold.fact}{a factor along which the cross-validation folds (\code{fold.col}) are specified. Factor \code{"1"} means the folds are specified for the rows of the main table.
Note that if the GP dimension is given by some real grouping factor (i.e. \code{p$GP_factor != "1"}), then the \code{fold.fact} must be that factor.}

\item{folds}{integer vector of folds to fit; if \code{NULL} (the default), all folds are fit}

\item{start.from.model}{an object of class \code{gp} to take the starting values from (from the \code{value} column in the hyperparameter table,
see \code{link[gpHyperparStartFromModel]}). If specified, then for each cross-validation fold model to be fit,
the starting values of the hyperparameters will be taken from the corresponding fold model wherever possible.}

\item{parallel}{should the cross-validation models be run in parallel? Uses foreach(), and requires parallel background for foreach() to be already registered.}

\item{fn.prefix}{character; prefix of file names for log files (\code{log.fn}) and dump files (\code{dump.fn}), in case these are not \code{NULL}}

\item{log.fn}{if not \code{NULL}, standard and error output of each fold model job will be saved into a log file with this file name. Special \code{\%} sequences can be used, see Details below.}

\item{dump.fn}{if not \code{NULL}, debug dump of given fold model will be saved upon an error, with this file name (the .rda extension will be added to it). Special \code{\%} sequences can be used, see Details below.}

\item{tr.max.lines}{the \code{max.lines} parameter for \code{traceback}, i.e. the maximum number of lines printed per call when error occurs}

\item{...}{options to be passed to the \code{\link[gpFit]{gpFit()}} method.}
}
\description{
Function will fit the GP model for each of the N cross-validation folds, and evaluate cross-validation prediction for the whole training data set.
}
\details{
The arguments \code{log.fn} and \code{dump.fn} allow for special sequences:
\itemize{
\item \code{\%f} - number of the cross-validation fold
\item \code{\%h} - hostname, i.e. the name of the machine where the worker job runs
\item \code{\%p} - process ID of the worker job
}
}
