% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpLGO.R
\name{gpLGO}
\alias{gpLGO}
\title{Calculate the approximation of the LGO-CV (leave-group-out cross-validation) and evaluate it.}
\usage{
gpLGO(
  gp,
  fold.col,
  fold.fact = "1",
  pred.null = NULL,
  lmFit.options = list(lik.hyperpar = "fix"),
  lmFit.pred.options = list(type = "terms"),
  hyperpar = gpHyperparList(gp)
)
}
\arguments{
\item{gp}{GP model object}

\item{fold.col}{either a name of a column in the main table, or a vector along factor \code{fold.fact}.
The column (or the supplied vector) must be a vector of integers from \code{1} to \code{N} (\code{N} being the number of folds),
specifying the number of the cross-validation fold the given record belongs to.}

\item{fold.fact}{a factor along which the cross-validation folds (\code{fold.col}) are specified. Factor \code{"1"} means the folds are specified for the rows of the main table.
Note that if the GP dimension is given by some real grouping factor (i.e. \code{gp$GP_factor != "1"}), then the \code{fold.fact} must be that factor.}

\item{pred.null}{optional predictions from the null model (\code{predict.lmFit(type = "terms")}). If these are available from the actual exact lmFitCV call,
it is better to provide them this way. Otherwise the null model predictions will be calculated just by fitting all data as training, which will result in a slightly
better null model (and a slightly lower pseudo-R2 statistics then).}

\item{lmFit.options}{list of options to be passed to \code{\link{lmFit}} when fitting the intercept-only null model. Only used when \code{pred.null = NULL}.}

\item{lmFit.pred.options}{list of options to be passed to the \code{predict.lmFit} method when predicting from the null model. Only used when \code{pred.null = NULL}.}

\item{hyperpar}{(optional) lists of hyperparameter values, as returned by \code{gpHyperparList()}; default is to use the hyperparameters from the \code{gp} object.}
}
\value{
the gp model object with approximate LGO-CV predictions and statistics in \code{gp$fit$LGOCV}.
}
\description{
Calculate the approximation of the LGO-CV (leave-group-out cross-validation) and evaluate it.
}
