% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpPack.R
\name{gpUnpack}
\alias{gpUnpack}
\title{Unpack the model - prepare it for computations (mostly predictions)}
\usage{
gpUnpack(
  gp,
  compute = TRUE,
  need.K = compute,
  need.L = FALSE,
  need.LTinv.rW = compute,
  num.correct.W.tol = 10 * sqrt(.Machine$double.eps)
)
}
\arguments{
\item{compute}{logical - should we prepare for computations (mostly predictions)? This is the CPU-heavy part. \code{compute = FALSE} will skip this.}

\item{need.K, need.L, need.LTinv.rW}{logical - in case of \code{compute = TRUE}, specify whether particular matrix (K, L) is needed. If only one of these matrices (K, L, need.LTinv.rW) is needed, it can
significantly save memory and sometimes also CPU time. But the object depends on each other (K -> L -> LTinv.rW),
so if you set this to \code{FALSE} for an object on which another needed object depends on,
you will only save the memory size of the resultant object, but not CPU time.

Note that the status of the \code{rW} object will be exactly the same as the \code{L} object (so it is directed by \code{need.L} argument).}

\item{num.correct.W.tol}{numerical correction of W. Applies only to the case of diagonal W (gp$W.type == "diag").
If not NULL, values of W will be corrected for small negative numbers, within given tolerance.}
}
\description{
Restore the auxiliary data structures, which are redundant and take significant space, but are needed for computations.
Note that restoring these structures can take some CPU time.
}
