% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpPack.R
\name{gpUnpack}
\alias{gpUnpack}
\title{Unpack the model - prepare it for computations (mostly predictions)}
\usage{
gpUnpack(gp, compute = TRUE, need.K = compute, need.L = compute)
}
\arguments{
\item{compute}{logical - should we prepare for computations (mostly predictions)? This is the CPU-heavy part. \code{compute = FALSE} will skip this.}

\item{need.K, need.L}{logical - in case of \code{compute = TRUE}, specify whether particular matrix (K, L) is needed. If only one of these matrices (K, L) is needed, it can
significantly save memory, and in case of L matrix, also CPU time. Note that if \code{compute = TRUE}, \code{K} matrix will be calculated regardless of \code{need.K};
so \code{need.K = FALSE} will only save the memory size of the resultant object, but not CPU time.}
}
\description{
Restore the auxiliary data structures, which are redundant and take significant space, but are needed for computations.
Note that restoring these structures can take some CPU time.
}
