% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmFit.R
\name{lmFit}
\alias{lmFit}
\title{Fit a linear version of a GP model}
\usage{
lmFit(
  gp,
  formula,
  data,
  table,
  beta.name = "beta_lm",
  lik.hyperpar = c("fix", "optimize"),
  lik.fix = NULL,
  lik.prefix = ".lik.",
  silent = TRUE,
  optCtrl = list(iter.max = 300, eval.max = 400)
)
}
\arguments{
\item{gp}{gp model object; its likelihood template will be used}

\item{formula}{the linear model formula}

\item{data}{object of class \code{gpData}; data used for the likelihood and the linear predictor}

\item{table}{character - name of the table in \code{data} that is going to be used to construct the model matrix for
the linear model}

\item{beta.name}{character; name of the linear predictor coefficient parameter}

\item{lik.hyperpar}{character; should likelihood hyperparameters be fixed or optimized?}

\item{lik.fix}{Can be either a single numeric value - then it will be used for all likelihood hyperparameters - or \code{NULL}. If \code{NULL}, values are taken from \code{gp$hyperpar$value}.}

\item{lik.prefix}{character; prefix added to likelihood hyperparameter names in \code{summary()}}

\item{silent}{logical; passed to \code{RTMB::MakeADFun}}

\item{optCtrl}{list; passed to \code{nlminb(control = )}}
}
\value{
Object of class \code{lmFit}
}
\description{
Fits a GP-free version of the model, where the latent Gaussian process \code{f}
is replaced by a linear predictor constructed from \code{formula}.
}
