% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{parallelJobWrapper}
\alias{parallelJobWrapper}
\title{Wrapper for parallel/background jobs}
\usage{
parallelJobWrapper(
  expr,
  working.dir = NULL,
  masterPID = NULL,
  log.fn = NULL,
  dump.fn = NULL,
  parallel = TRUE,
  tr.max.lines = 5,
  log.append = FALSE
)
}
\arguments{
\item{expr}{expression - the job code to be run}

\item{working.dir}{working directory}

\item{masterPID}{the PID (process ID) of the process that is dispatching this parallel computation}

\item{log.fn}{if not \code{NULL}, standard and error output of the job will be saved into a log file with this file name. Special \code{\%} sequences can be used, e.g. \code{"log-\%h_\%p.txt"}, see Details below.}

\item{dump.fn}{if not \code{NULL}, debug dump will be saved upon an error, with this file name (the .rda extension will be added to it). Special \code{\%} sequences can be used, e.g. \code{"dump-\%h_\%p"}, see Details below.}

\item{parallel}{should be set to \code{TRUE} for the wrapper to work normally. The \code{FALSE} setting is just for some perhaps
debugging purposes, when we want to run the job in non-parallel, interactive mode, in which case most of the wrapper function (logging, setting working directory)
will be disabled.}

\item{tr.max.lines}{the \code{max.lines} parameter for \code{traceback}, i.e. the maximum number of lines printed per call when error occurs}

\item{log.append}{logical; if \code{TRUE}, append to \code{log.fn} instead of overwriting it}
}
\description{
Useful wrapper for parallel jobs, run by e.g. foreach(), or, in general, for any non-interactive (background) jobs.
It will make sure the working directory is set correctly, standard and error outputs are logged, and in case of
error, a dump file will be saved for debugging purposes.
}
\details{
The arguments \code{log.fn} and \code{dump.fn} allow for special sequences:
\itemize{
\item \code{\%h} - hostname, i.e. the name of the machine where the worker job runs
\item \code{\%p} - process ID of the worker job
}
}
