% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred.R
\name{pred}
\alias{pred}
\title{Internal predict function}
\usage{
pred(
  gp,
  predx,
  same = FALSE,
  hyperpar = gpHyperparList(gp),
  components = NULL,
  comp_missing = c("avg", "none"),
  w = NULL,
  groupMeans = NULL,
  se.fit = TRUE,
  cov.fit = FALSE,
  maxn = 250,
  Kx.cache = NULL,
  Kxx.cache = NULL,
  recursive = FALSE
)
}
\arguments{
\item{components}{character vector of names of the components of covariance that should be used for this prediction}

\item{se.fit}{return the standard errors of the predicted values?}

\item{cov.fit}{return the covariance matrix of the predicted values?}
}
\description{
Internal predict function
}
