% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.gp.R
\name{predict.gp}
\alias{predict.gp}
\title{Predict from a fitted Gaussian process model}
\usage{
\method{predict}{gp}(
  gp,
  newdata = NULL,
  type = c("latent", "terms", "response"),
  se.fit = FALSE,
  cov.fit = FALSE,
  conf.int = FALSE,
  conf.level = 0.95,
  components = NULL,
  comp_missing = c("avg", "none"),
  w = NULL,
  groupMeans = NULL,
  hyperpar = gpHyperparList(gp),
  maxn = Inf,
  pred.sims = 1e+05,
  Kx.cache = NULL,
  Kxx.cache = NULL,
  ...
)
}
\arguments{
\item{gp}{object of class gp, the fitted Gaussian process model}

\item{newdata}{object of class gpData, \emph{\strong{unscaled}} (i.e. not passed through \code{gpDataPrepare()}) data! They must be on the same scale as
the unscaled training dataset used for fitting the model. If \code{NULL} (the default), predictions will be made on the training dataset (which
is going to be super fast if conf.int = FALSE)}

\item{type}{character; type of prediction:
\describe{
\item{\code{"latent"}}{predictions directly from the Gaussian Process itself (\code{f} and optionally \code{f_SE}), always in the dimension of the GP (\code{gp$GP_factor})}
\item{\code{"terms"}}{predictions passed through first stage of the likelihood template, that is, including GP (\code{f} and optionally \code{f_SE}) along with all the other terms used in the likelihood template. If \code{gp$lik.reindex2main} is \code{TRUE},
the predictions will be reindexed to the dimension of the main table.}
\item{\code{"response"}}{predictions passed through the first two stages of the likelihood template, that is, after applying the inverse link function. If \code{gp$lik.reindex2main} is \code{TRUE},
the predictions will be reindexed to the dimension of the main table.}
}}

\item{cov.fit}{logical; if \code{TRUE}, the covariance matrix of the predictions will also be returned (can be very memory consuming for large datasets!).
Default is \code{FALSE}.}

\item{conf.int}{logical; should confidence intervals be calculated? If \code{FALSE}, no confidence intervals will be calculated,
which makes the prediction much faster (but just in case both \code{se.fit} and \code{cov.fit} are also \code{FALSE}).
If \code{TRUE}, \code{se.fit} will be automatically set to \code{TRUE}, since the standard errors are needed for the CI calculation.}

\item{conf.level}{numeric; confidence level for the confidence intervals. Default is 0.95.}

\item{components}{of the covariance formula. One of the following:
\itemize{
\item NULL : all components will be selected (the default)
\item character vector: components to be selected
\item formula: additive formula specifying components, same syntax as in \link[stats:update.formula]{update.formula}, e.g.
\tabular{ll}{
\code{~.}          \verb{   } \tab all components \cr
\code{~.-env-spat} \verb{   } \tab all components except \code{env} and \code{spat} \cr
\code{~spat+year}  \verb{   } \tab only components \code{spat} and \code{year} \cr
}
}}

\item{comp_missing}{how to compensate for missing components (if there are any), in mean and variance of the predictions:
\describe{
\item{\code{"none"}}{no compensation. In this case, the interpretation of the result's variance isn't much clear (!!!??? why?).}
\item{\code{"avg"}}{the missing components are replaced by "averaging over them" - more
precisely, it predicts for the average effects of these components - for these, the variance/CI corresponds to confidence interval
(uncertainty around the mean predictions) and not the prediction interval - see http://www.sthda.com/english/articles/40-regression-analysis/166-predict-in-r-model-predictions-and-confidence-intervals/
For models which have grouping factors, be aware of how the average is weighted - see CAUTION CAU001 in the code.
}
}
Thus, the most reasonable setting now is \code{"avg"}; unless one cares only about the prediction mean (\code{f}) of the latent variable, in which case
a simple "none" would also do the job.}

\item{w}{providing this optional argument will calculate average prediction(s) of all data points in this dataset, given these weights. Can
be either a vector of the same length as the size of \code{newdata} (in which case a single average is calculated), or,
if more different averages with different weights are desired, \code{w} can be a matrix where each column corresponds to one average - one set of weights
(i.e. number of rows of \code{w} must be equal to the size of \code{newdata}, number of columns equal to the number of different averages to be calculated).}

\item{groupMeans}{optional, factor of the size of the predicted dataset. If provided, mean predictions for each factor level are calculated.}

\item{hyperpar}{(optional) lists of hyperparameter values, as returned by \code{gpHyperparList()}; default is to use the hyperparameters from the \code{gp} object.}

\item{maxn}{maximum dataset size (along the \code{gp$GP_factor}) to fit at once; if the dataset is larger, it will be split into chunks of size \code{maxn}
before predictions are calculated. Use \code{Inf} to disable the splitting entirely. Choosing a suitable value may speed up the computation when \code{se.fit = TRUE} and \code{cov.fit = FALSE},
because the Kxx matrices are then smaller (that's probably the only part where splitting helps) - but testing is needed. (Note: splitting disabled for now
by default, since in some models we would need to specify along which factor to split - cannot always use gp$GP_factor)}

\item{pred.sims}{(currently unused) number of simulations to be used for calculating the response scale predictions and their CIs. Default is 100000.
Higher values give more accurate results, but are slower. Ignored if \code{conf.int = FALSE}.}

\item{Kx.cache}{optional, object returned by \code{K_cache()} function, to speed up repeated calls to \code{predict.gp()} with the same \code{newdata} by
caching parts of the K(training_data, newdata) matrix.}

\item{Kxx.cache}{optional, object returned by \code{K_cache()} function, to speed up repeated calls to \code{predict.gp()} with the same \code{newdata} by
caching parts of the K(newdata, newdata) matrix.}

\item{link}{character, optional. In case of \code{type = "response"}, what should be the link function to take inverse of for calculating the derived quantity (response scale)? Character as passed to \code{make.link()}.}

\item{parname}{character, optional. In case of \code{type = "response"}, what should be the name of the parameter which is the derived quantity on the response scale?}
}
\value{
A matrix of predictions, with columns \code{f} and \code{f_SE} (if \code{se.fit = TRUE}). If \code{cov.fit = TRUE}, returns a named list, \code{pred} will be
the mentioned matrix and \code{cov} will be the full covariance matrix. For the dimension of the prediction, see the \code{type} parameter.
}
\description{
Predict from a fitted Gaussian process model.
}
