% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{simulate_lifecycle}
\alias{simulate_lifecycle}
\title{Simulate consumption / savings lifecycle problem}
\usage{
simulate_lifecycle(
  N,
  T,
  x_grid,
  N_shock,
  U_shock,
  P,
  init_x,
  G,
  v,
  sigma_n,
  sigma_u,
  gamma_0,
  gamma_1,
  R,
  p_noinc,
  beta,
  rho
)
}
\arguments{
\item{N}{number of simulations}

\item{T}{number of time periods}

\item{x_grid}{grid of permanent income we use for consumption rule}

\item{N_shock}{matrix of permanent income shocks}

\item{U_shock}{matrix of temporary income shocks}

\item{G}{growth of permanent income}

\item{v}{time-varying family size}

\item{sigma_n}{sd of permanent income}

\item{sigma_u}{sd temporary income shocks}

\item{R}{gross return on assets}

\item{p_noinc}{probability that income is zero}

\item{beta}{time discount factor}

\item{rho}{CRRA risk aversion}

\item{mu_a}{average log starting assets}

\item{sigma_a}{sd of log starting assets}
}
\description{
Simulate consumption / savings lifecycle problem
}
